% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AAV.R
\name{AAV}
\alias{AAV}
\title{Average Annual Variabililty (in Catch)}
\usage{
AAV(df)
}
\arguments{
\item{df}{A \link{data.frame}}
}
\value{
A vector of the AAV values
}
\description{
Average Annual Variabililty (in Catch)
}
