% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-append-objs-to-list.R
\name{append_objs_to_list}
\alias{append_objs_to_list}
\title{Append objects to list}
\usage{
append_objs_to_list(lst = NULL, ...)
}
\arguments{
\item{lst}{The list to append to}

\item{...}{The objects to append}
}
\value{
The modified list
}
\description{
Append objects to list
}
