% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{apply_quantiles}
\alias{apply_quantiles}
\title{Apply arbitrary quantiles to a column of data}
\usage{
apply_quantiles(df = NULL, col = NULL, probs = c(0.05, 0.5, 0.95))
}
\arguments{
\item{df}{A \link{data.frame}}

\item{col}{A column name found in \code{df}}

\item{probs}{A vector of probabilities to be passed to the \code{\link[stats:quantile]{stats::quantile()}} function}
}
\value{
A \link{data.frame} with a new column for wach value in the \code{probs} vector
}
\description{
Apply arbitrary quantiles to a column of data
}
