% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_catch_seas_country.R
\name{calc_catch_seas_country}
\alias{calc_catch_seas_country}
\title{Calculate the catch proportions by season and country as an average of the last N years}
\usage{
calc_catch_seas_country(
  data_csv_dir = NULL,
  n_yrs = 10,
  weight = c(1, 1, 1, 1)
)
}
\arguments{
\item{data_csv_dir}{The directory in which the assessment csv files containing catch by month are located}

\item{n_yrs}{The number of year into the past to use in the calculations}

\item{weight}{A vector of 4 for the weights given to each season}
}
\value{
A data frame with two rows, one for each country, and 4 columns, one for
each season. Canada will be the first row and the US the second
}
\description{
Calculate the catch proportions by season and country as an average of the last N years
}
