% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ss-model-loading.R
\name{calc_mcmc}
\alias{calc_mcmc}
\title{Return a list of mcmc calculations, e.g. quantiles for various values}
\usage{
calc_mcmc(
  mcmc,
  ss_mcmc_quants = NULL,
  biomass_scale = 2e+06,
  recruitment_scale = 1e+06,
  ...
)
}
\arguments{
\item{mcmc}{The output of the \code{\link[r4ss:SSgetMCMC]{r4ss::SSgetMCMC()}} function as a data.frame}

\item{ss_mcmc_quants}{Quantile probability value. See the \code{probs} argument in \code{\link[stats:quantile]{stats::quantile()}}.
The vector must have 3 values and be in ascending order}

\item{biomass_scale}{Scale the biomass by this amount. The default is 2e6 because
biomass will be shown in the millions of tonnes and it is female only}

\item{recruitment_scale}{Scale the recruitment by this amount. The default is 1e6
because recruitment will be shown in millions of tonnes}

\item{...}{Absorb arguments destined for other functions}
}
\value{
A \link{list} of MCMC calculated values
}
\description{
Return a list of mcmc calculations, e.g. quantiles for various values
}
