% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-calc-mean-age.R
\name{calc_mean_age}
\alias{calc_mean_age}
\title{Calculate mean age in the population}
\usage{
calc_mean_age(agemat = NULL, maxage = NULL)
}
\arguments{
\item{agemat}{matrix of ages}

\item{maxage}{maximum age to include in the calculation}
}
\value{
A vector of the mean ages by year, with names attribute as
year names (character)
}
\description{
Calculate mean age in the population
}
