% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-calc-mean-waa.R
\name{calc_mean_waa}
\alias{calc_mean_waa}
\title{Calculate the mean weights-at-age for all rows in a data frame}
\usage{
calc_mean_waa(d, yr_col = "Yr")
}
\arguments{
\item{yr_col}{Name of the column to use for the year. It must be present
in the \code{waa} data frame or an error will occur}

\item{waa}{A weigth-at-age data frame}
}
\value{
A single-row data frame containing the averages
}
\description{
Calculate the mean weights-at-age for all rows in a data frame
}
