% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{calc_quantiles}
\alias{calc_quantiles}
\alias{calc_term_quantiles}
\title{Calculate and insert columns containing arbitrary quantiles for a particular column}
\usage{
calc_quantiles(
  df = NULL,
  col = NULL,
  probs = c(0.05, 0.25, 0.5, 0.75, 0.95),
  include_mean = TRUE
)

calc_term_quantiles(
  df = NULL,
  grp_col = NULL,
  col = NULL,
  min_yr = NA,
  max_yr = NA,
  probs = c(0.05, 0.25, 0.5, 0.75, 0.95),
  mean_multiplier = 1
)
}
\arguments{
\item{df}{A \link{data.frame} with columns \code{year}, the string assigned to \code{grp_col}, and
the string assigned to \code{col}}

\item{col}{A column name on which to perform the calculations. Must be in \code{df} or an error
will be thrown}

\item{probs}{A vector of quantiles to pass to \code{\link[stats:quantile]{stats::quantile()}}}

\item{include_mean}{If TRUE, include the mean in the output}

\item{grp_col}{The column name to use for grouping the data}

\item{min_yr}{The minimum year in the term. If \code{NA}, the minimum value in the \link{data.frame}
will be used}

\item{max_yr}{The maximum year in the term. If \code{NA}, the maximum value in the \link{data.frame}
will be used}

\item{mean_multiplier}{This value will be multiplied by the mean of the values
before the quantiles are calculated}
}
\value{
A \link{data.frame} with a new column for each value in the \code{probs} vector

A single-row \link{data.frame} containing the quantile values of the run means
}
\description{
Calculate and insert columns containing arbitrary quantiles for a particular column

Calculate quantiles for the mean values (across years) of each value in
the group given by \code{grp_col} for a term of years
}
\examples{
library(tibble)
library(dplyr)
library(purrr)
pq <- tribble(
  ~year, ~grp, ~val,
  2000,    1,  2.1,
  2001,    1,  3.4,
  2002,    1,  4.5,
  2003,    1,  5.6,
  2004,    1,  6.7,
  2000,    2,  3.1,
  2001,    2,  4.4,
  2002,    2,  5.5,
  2003,    2,  6.6,
  2004,    2,  8.7,
  2000,    3, 13.1,
  2001,    3, 14.4,
  2002,    3, 15.5,
  2003,    3, 16.6,
  2004,    3, 18.7)

probs <- c(0.05, 0.25, 0.5, 0.75, 0.95)

yrs <- sort(unique(pq$year))
df <- pq \%>\%
  group_by(year) \%>\%
  group_map(~ calc_quantiles(.x, col = "val", probs = probs)) \%>\%
  map_df(~{.x}) \%>\%
  mutate(year = yrs) \%>\%
  select(year, everything())
library(tibble)
library(dplyr)
library(purrr)
pq <- tribble(
  ~year, ~grp, ~val,
  2000,    1,  2.1,
  2001,    1,  3.4,
  2002,    1,  4.5,
  2003,    1,  5.6,
  2004,    1,  6.7,
  2000,    2,  3.1,
  2001,    2,  4.4,
  2002,    2,  5.5,
  2003,    2,  6.6,
  2004,    2,  8.7,
  2000,    3, 13.1,
  2001,    3, 14.4,
  2002,    3, 15.5,
  2003,    3, 16.6,
  2004,    3, 18.7)

probs <- c(0.05, 0.25, 0.5, 0.75, 0.95)

j <- calc_term_quantiles(pq,
                         grp_col = "grp",
                         col = "val",
                         probs = probs)
}
