% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcSE.R
\name{calc_standard_error_ssb}
\alias{calc_standard_error_ssb}
\title{Calculate the standard error on SSB between the Operating Model and the Estimation Model
for all runs in the list}
\usage{
calc_standard_error_ssb(lst)
}
\arguments{
\item{lst}{list of MSE results}
}
\value{
A \link{data.frame} of 3 columns: The run number, the standard error of spawning
stock biomass from EM in regards to the OM, and the year
}
\description{
Calculate the standard error on SSB between the Operating Model and the Estimation Model
for all runs in the list
}
