% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tac_estimated_vs_realized.R
\name{calc_tac_est_vs_real}
\alias{calc_tac_est_vs_real}
\title{Calculated the estimated TAC vs the Realized TAC (from data)}
\usage{
calc_tac_est_vs_real()
}
\value{
A list of 2 items. (1) A \link{data.frame} as read in from the TAC.csv data file
found in this package, and (2) A \link{data.frame} as calculated by this function
}
\description{
Calculated the estimated TAC vs the Realized TAC (from data)
}
\details{
SSB is a vector of equally spaced values between zero and double the
Operating model SSB0 value (for a single-space, single-season model). TAC is
zero for SSB / SSB0 less than 0.1, 0.4\ * SSB for values greater than 0.4, and
for values between 0.1 and 0.4, TAC is 0.4\* SSB \* (SSB - 0.1SSB0) \*
the ratio (0.4 \* SBB0 / SSB) / (0.4 \* SSB0 - 0.1SSB0)
}
