% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcF.R
\name{calcF}
\alias{calcF}
\title{Title}
\usage{
calcF(Fmethod = "hybrid", Etemp, Btmp, N, weight, M, Fsel)
}
\arguments{
\item{Fmethod}{hybrid only viable option right now}

\item{Etemp}{Catch input}

\item{Btmp}{estimated available biomass}

\item{N}{numbers at age}

\item{weight}{weight at age}

\item{M}{natural mortality}

\item{Fsel}{fishing selectivity}
}
\value{

}
\description{
Title
}
\examples{
calcF(Fmethod = 'hybrid') # Not a working example
}
