% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcMeanAge.R
\name{calcMeanAge}
\alias{calcMeanAge}
\title{Calculate mean age in the population}
\usage{
calcMeanAge(agemat, maxage)
}
\arguments{
\item{agemat}{matrix of ages}

\item{maxage}{maximum age to include in the calculation}
}
\value{
A vector of the mean ages
}
\description{
Calculate mean age in the population
}
