% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcSE.R
\name{calcSE}
\alias{calcSE}
\title{Calculate the error on SSB}
\usage{
calcSE(ls.MSE)
}
\arguments{
\item{ls.MSE}{list of MSE results}
}
\value{
The standard error of spawning stock biomass from EM
in regards to the OM
}
\description{
Calculate the error on SSB
}
