% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Check_Identifiable_vs2.R
\name{Check_Identifiable_vs2}
\alias{Check_Identifiable_vs2}
\title{Calculate the estimated parameters with uncertainty}
\usage{
Check_Identifiable_vs2(obj)
}
\arguments{
\item{obj}{\link{TMB} object}
}
\value{
Return a list of estimated parameters and uncertainty
}
\description{
Calculate the estimated parameters with uncertainty
}
