% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{col2hex}
\alias{col2hex}
\title{Convert color name to hex string. Code borrowed from \code{\link[gplots:col2hex]{gplots::col2hex()}}}
\usage{
col2hex(cname)
}
\arguments{
\item{cname}{Color name as a common name. If a hex string is supplied, it will be returned unchanged.}
}
\value{
The HEX string representing the color
}
\description{
Convert color name to hex string. Code borrowed from \code{\link[gplots:col2hex]{gplots::col2hex()}}
}
