% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{color_facet_backgrounds}
\alias{color_facet_backgrounds}
\title{Color the backgrounds of the facet labels in a ggplot object}
\usage{
color_facet_backgrounds(
  g = NULL,
  facet_back_cols = NULL,
  facet_back_alpha = 99
)
}
\arguments{
\item{g}{The ggplot object}

\item{facet_back_cols}{A vector of the colors to apply to the facet backgrounds}

\item{facet_back_alpha}{transparency between 0 and 99}
}
\value{
The modified ggplot object
}
\description{
Color the backgrounds of the facet labels in a ggplot object
}
