% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_om.R
\name{compare_age_in_catch}
\alias{compare_age_in_catch}
\title{Plot the average age in catch for the OM}
\usage{
compare_age_in_catch(obj)
}
\arguments{
\item{obj}{One of the OM list elements "catch_save_age" or "catch_n_save_age"}
}
\value{
a \code{\link[ggplot2:ggplot]{ggplot2::ggplot()}} object
}
\description{
Plot the average age in catch for the OM
}
