% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_om.R
\name{compare_biomass}
\alias{compare_biomass}
\title{Plot a comparison biomass (SSB) from the OM and the old OM, which is read in as an rds object}
\usage{
compare_biomass(
  om,
  fn = "C:/github/pacific-hake/pacifichakemse/tests/testthat/ssb_out.rds"
)
}
\arguments{
\item{om}{Output from the \code{\link[=run_om]{run_om()}} function}

\item{fn}{Location of the om_output.rds file from the old code"s base OM run}
}
\value{
a \code{\link[ggplot2:ggplot]{ggplot2::ggplot()}} object
}
\description{
Plot a comparison biomass (SSB) from the OM and the old OM, which is read in as an rds object
}
