% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_om.R
\name{compare_move}
\alias{compare_move}
\title{Plot movement into each country by season. To use:
p <- compare_move(om)
do.call("grid.arrange", c(p, ncol = 2))}
\usage{
compare_move(om)
}
\arguments{
\item{om}{Output from the \code{\link[=run_om]{run_om()}} function}
}
\value{
a list of 4 \code{\link[ggplot2:ggplot]{ggplot2::ggplot()}} objects
}
\description{
Plot movement into each country by season. To use:
p <- compare_move(om)
do.call("grid.arrange", c(p, ncol = 2))
}
