% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{compare_tmb_data}
\alias{compare_tmb_data}
\title{Compare two data object inputs for input into \code{pacifichakemse.cpp}}
\usage{
compare_tmb_data(d1, d2, p1, p2)
}
\arguments{
\item{d1}{Data object 1}

\item{d2}{Data object 2}

\item{p1}{Parameter object 1}

\item{p2}{Parameter object 2}
}
\description{
Used while developing the code to compare old and new data being input.
They need to be exactly the same
}
