% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{conv_d}
\alias{conv_d}
\title{Read in and convert data objects from the
old code to make sure new code produced the same likelihoods}
\usage{
conv_d(yr)
}
\arguments{
\item{yr}{The year for file read (file will be d2018.rds for example)}
}
\value{
The data with converted names
}
\description{
Read in and convert data objects from the
old code to make sure new code produced the same likelihoods
}
