% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{conv_p}
\alias{conv_p}
\title{Read in and convert parameter objects from the
old code to make sure new code produced the same likelihoods}
\usage{
conv_p(yr)
}
\arguments{
\item{yr}{The year for file read (file will be p2018.rds for example)}
}
\value{
The parameters with converted names
}
\description{
Read in and convert parameter objects from the
old code to make sure new code produced the same likelihoods
}
