% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{conv_vec_to_mse_df}
\alias{conv_vec_to_mse_df}
\title{Convert a vector of aggregated values into a \link{data.frame} containing
the year, run, value, and possibly country as columns}
\usage{
conv_vec_to_mse_df(
  vec = NULL,
  col = NULL,
  yr_vec = NULL,
  cntry = NULL,
  probs = c(0.05, 0.25, 0.5, 0.75, 0.95)
)
}
\arguments{
\item{vec}{The aggregated vector of data. This must be the same length as \code{run_vec}
and \code{yr_vec}}

\item{col}{The name of the column to create in teh new data frame}

\item{yr_vec}{The vector of years to place in the \link{data.frame}. This must be the
same length as \code{vec}}

\item{cntry}{The name for the country. If \code{NULL}, no country column will be added}

\item{probs}{A vector of quantiles to pass to \code{\link[stats:quantile]{stats::quantile()}}}
}
\value{
A \link{data.frame} containing the year, run, value, and possibly country as columns
}
\description{
Convert a vector of aggregated values into a \link{data.frame} containing
the year, run, value, and possibly country as columns
}
