% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create-plot-objects.R
\name{create_plot_objects}
\alias{create_plot_objects}
\title{Create and/or load RDS file contents from multiple MSE runs}
\usage{
create_plot_objects(
  scenarios = c("biasadjust", "climate", "hcr", "selectivity"),
  om_only = rep(FALSE, length(scenarios)),
  overwrite_rds = TRUE,
  ...
)
}
\arguments{
\item{scenarios}{A vector of directory names inside the results
directory with MSE results}

\item{om_only}{A vector of logicals of the same length as \code{scenarios}
indicating if the run is an OM only run (TRUE) or a full MSE with EM
estimation (FALSE)}

\item{overwrite_rds}{If TRUE, all RDS files will be re-created (takes
a long time). If FALSE, the files that exists will be loaded and returned}

\item{...}{Arguments to pass to \code{\link[=load_mse_plot_data]{load_mse_plot_data()}}}
}
\value{
A list of length of \code{scenarios} containing MSE output data
}
\description{
Create and/or load RDS file contents from multiple MSE runs
}
