% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{csv_data}
\alias{csv_data}
\title{Load data from CSV files}
\usage{
csv_data(sel_hist = TRUE)
}
\arguments{
\item{sel_hist}{Logical Load the PSEL file?}
}
\value{
A list of the loaded data.frames
}
\description{
Load data from CSV files
}
