% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-df-identical.R
\name{df_identical}
\alias{df_identical}
\title{Are matrices identical}
\usage{
df_identical(wa1, wa2, nm_wa1, nm_wa2, diff_tol = 1e-20)
}
\arguments{
\item{wa1}{waa matrix 1}

\item{wa2}{waa matrix 2}

\item{nm_wa1}{name of waa matrix 1}

\item{nm_wa2}{name of waa matrix 2}

\item{diff_tol}{Tolerance in the numbers for identical}
}
\value{
\code{TRUE} or \code{FALSE}
}
\description{
Are matrices identical
}
