% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df_lists.R
\name{df_lists}
\alias{df_lists}
\title{A list of values}
\usage{
df_lists(ls.save, nms)
}
\arguments{
\item{ls.save}{list of MSE data}

\item{nms}{names of runs}
}
\value{
A list of values
}
\description{
A list of values
}
