% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df_lists_OM.R
\name{df_lists_OM}
\alias{df_lists_OM}
\title{A list of values for the Operating Model}
\usage{
df_lists_OM(ls.save, nms)
}
\arguments{
\item{ls.save}{MSE results}

\item{nms}{names to save in data frame}
}
\value{
A list of values for the Operating Model
}
\description{
A list of values for the Operating Model
}
