% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/on_unload.R
\name{.onUnload}
\alias{.onUnload}
\title{Unload the dynamic library when the package is unloaded}
\usage{
.onUnload(libpath)
}
\arguments{
\item{libpath}{Path where the library is located}
}
\value{
Nothing
}
\description{
Unload the dynamic library when the package is unloaded
}
