% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{extract_age_comps}
\alias{extract_age_comps}
\title{Extract the age comps from the SS assessment model output into a format
required for input into the TMB assessment model}
\usage{
extract_age_comps(
  ss_model = NULL,
  fleet = 1,
  s_yr = NULL,
  m_yr = NULL,
  fill = -1,
  yr_col = "Yr",
  fleet_col = "FltSvy"
)
}
\arguments{
\item{ss_model}{SS model input/output as read in by \code{\link[=load_ss_model_from_rds]{load_ss_model_from_rds()}}}

\item{fleet}{1 for fishery, 2 for survey}

\item{s_yr}{See \code{\link[=load_data_seasons]{load_data_seasons()}}}

\item{m_yr}{See \code{\link[=load_data_seasons]{load_data_seasons()}}}

\item{fill}{Value to replace NAs in the table, it can also be NA, but not NULL}

\item{yr_col}{The name of the column in \code{ss_model$dat$agecomp} that contains
the year}

\item{fleet_col}{The name of the column in \code{ss_model$dat$agecomp} that contains
the fleet code}
}
\value{
The matrix representing all years from \code{s_yr} to \code{m_yr} with data included
for years which are found in the \code{ss_model} data. Years not in the \code{ss_model} data
will be filled with the value of \code{fill}
}
\description{
Extract the age comps from the SS assessment model output into a format
required for input into the TMB assessment model
}
\details{
Proportions at age for each year are calculated and returned
}
