% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Check_Identifiable_vs2.R
\name{extract_fixed}
\alias{extract_fixed}
\title{Title}
\usage{
extract_fixed(obj)
}
\arguments{
\item{obj}{TMB object}
}
\value{
Return estimated parameters and uncertainty
}
\description{
Title
}
\examples{



}
