% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-extract-initial-values.R
\name{extract_initial_values}
\alias{extract_initial_values}
\title{TMB helper function. Extract parameter from \code{obj} to use as initial
values in optimization}
\usage{
extract_initial_values(obj)
}
\arguments{
\item{obj}{Object to be used as input to \code{\link[TMB:MakeADFun]{TMB::MakeADFun()}}}
}
\value{
A vector of initial values of parameters
}
\description{
TMB helper function. Extract parameter from \code{obj} to use as initial
values in optimization
}
\details{
The code was copied from a non-exported function in the
\code{TMBhelper} package called \code{extract_fixed()}
}
