% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-extract-params-tab.R
\name{extract_params_tmb}
\alias{extract_params_tmb}
\title{Extract parameter estimates from the linear par vector returned by TMB
to a list with parameter names and values}
\usage{
extract_params_tmb(obj)
}
\arguments{
\item{obj}{The output from the TMB model}
}
\value{
A named \link{list} of parameter estimates extracted from \code{obj$par}
}
\description{
Extract parameter estimates from the linear par vector returned by TMB
to a list with parameter names and values
}
