% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/url-file-exists.R
\name{files_exist}
\alias{files_exist}
\title{Test whether or not files exist. They can be URLs or regular files}
\usage{
files_exist(fns_lst)
}
\arguments{
\item{fns_vec}{A character vector of URLs or regular full path filenames.
The function \code{\link[base:files]{base::file.exists()}} will be used for filenames not
containing 'https://' or 'http://'}
}
\value{
A logical vector of \code{TRUE} and \code{FALSE} corresponding to
the \code{fns_lst} file
locations
}
\description{
Test whether or not files exist. They can be URLs or regular files
}
