% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-fill-waa-years.R
\name{fill_waa_years}
\alias{fill_waa_years}
\title{Fill in missing year rows in a weight-at-age data frame with supplied values}
\usage{
fill_waa_years(waa, values, yr_col = "Yr", yrs = NULL)
}
\arguments{
\item{waa}{A weight-at-age data frame}

\item{values}{A single-row data frame as output by \code{\link[=calc_mean_waa]{calc_mean_waa()}}}

\item{yr_col}{Name of the column to use for the year. It must be present
in the \code{waa} data frame or an error will occur}

\item{yrs}{The years to include in the calculation. If \code{NULL}, all years
in \code{waa} will be used. Negatives will be removed from the calculation. For
example if yrs = -c(1966, 1967) then all years except 1966 and 1967 will
be used in the calculation}
}
\value{
A data frame, the same as the input \code{waa} data frame, but with
the missing years defined by \code{yrs} filled in with the mean data found in
\code{means}
}
\description{
Fill in missing year rows in a weight-at-age data frame with supplied values
}
