% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{fn_finder}
\alias{fn_finder}
\title{Get a calling function's name from within the function}
\usage{
fn_finder(
  skip_frames = 1,
  skip_names = "(FUN)|(.+apply)|(replicate)",
  ret_stack = FALSE,
  extra_perf_per_level = "\\t"
)
}
\arguments{
\item{skip_frames}{The level in the calling stack to look. 1 is in the current
function, 2 is one before, etc.}

\item{extra_perf_per_level}{}
}
\value{

}
\description{
Get a calling function's name from within the function
}
