% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_rerun_OM.R
\name{fn_rerun_OM}
\alias{fn_rerun_OM}
\title{Rerun an OM from an MSE by using catch as input}
\usage{
fn_rerun_OM(folders, df)
}
\arguments{
\item{folders}{TODO}

\item{df}{TODO}
}
\value{
Nothing
}
\description{
Rerun an OM from an MSE by using catch as input
}
