% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fnMSE.R
\name{fnMSE}
\alias{fnMSE}
\title{Initialize the operating model}
\usage{
fnMSE(df, simyears = 30, TAC = 1, seeds = NA, nruns = 100)
}
\arguments{
\item{df}{parameters}

\item{simyears}{number of years to simulate}

\item{TAC}{which harvest control rule to use}

\item{seeds}{seed}

\item{nruns}{number of iterations}
}
\value{
QA list of length 2: (1) is a list of number of runs of MSE results,
output by \code{\link[=run_multiple_MSEs]{run_multiple_MSEs()}}, (2) is a simulation object as output from
\code{\link[=run.agebased.true.catch]{run.agebased.true.catch()}}
}
\description{
Initialize the operating model
}
\examples{
\dontrun{
fnMSE(df, nruns = 15) # Run MSE with 15 iterations
}
}
