% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df_MSE.R
\name{fnsum}
\alias{fnsum}
\title{sum spatial and seasonal data into a data frame}
\usage{
fnsum(x, idx)
}
\arguments{
\item{x}{list of data}

\item{idx}{which index to sum over}
}
\value{
returns list as data frame
}
\description{
sum spatial and seasonal data into a data frame
}
\examples{
catchtmp.country <- map(df.MSE[names(df.MSE) == 'Catch'],.f = fnsum, idx = c(2,3))


}
