% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{format_wage_df}
\alias{format_wage_df}
\title{Extract the weight-at-age data for the given Fleet}
\usage{
format_wage_df(df = NULL, fleet = NULL)
}
\arguments{
\item{df}{A weight-at-age \link{data.frame}}

\item{fleet}{The fleet number to extract}
}
\value{
A \link{data.frame}
}
\description{
Extract the weight-at-age data for the given Fleet
}
