% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{format_wage_matrix}
\alias{format_wage_matrix}
\title{Convert the output of \code{\link[=format_wage_df]{format_wage_df()}} into a matrix of correct dimensions
for the TMB input}
\usage{
format_wage_matrix(df = NULL)
}
\arguments{
\item{df}{A weight-at-age \link{data.frame} as output from \code{\link[=format_wage_df]{format_wage_df()}}}
}
\value{
A \link{matrix}
}
\description{
Convert the output of \code{\link[=format_wage_df]{format_wage_df()}} into a matrix of correct dimensions
for the TMB input
}
