% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{func_name}
\alias{func_name}
\title{Returns a calling function's name \code{levels_up} levels up the stack trace}
\usage{
func_name(levels_up = 1)
}
\arguments{
\item{levels_up}{How many levels back in the stack trace to look for the
function name}
}
\value{
A calling function's name \code{levels_up} levels up the stack trace
}
\description{
Returns a calling function's name \code{levels_up} levels up the stack trace
}
\examples{
f <- function(){
  message("You are in ", func_name())
}
}
