% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{get_age_dat}
\alias{get_age_dat}
\title{Extract row(s) of age data from a \link{data.frame}}
\usage{
get_age_dat(d = NULL, yr = NULL)
}
\arguments{
\item{d}{A \link{data.frame} in the format of the \verb{wage_*} data frames as
output by \code{\link[=load_data_seasons]{load_data_seasons()}}}

\item{yr}{A vector of years to extract row of data for}
}
\value{
A data frame with the rows requested
}
\description{
Extract row(s) of age data from a \link{data.frame}
}
