% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_em_outputs.R
\name{get_em_outputs}
\alias{get_em_outputs}
\title{Get outputs from the Estimation models}
\usage{
get_em_outputs(em, quants = c(0.05, 0.5, 0.95), ...)
}
\arguments{
\item{em}{The esimation model object for a single scenario}

\item{quants}{A vector of quantiles to apply}

\item{...}{Absorbs unused arguments}
}
\value{
A list of the EM outputs
}
\description{
Get outputs from the Estimation models
}
