% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getF.R
\name{get_f}
\alias{get_f}
\title{Calculate the fishing mortality based on catch}
\usage{
get_f(e_temp, b_temp, m_season, f_sel, n_temp, w_catch, method = "Pope")
}
\arguments{
\item{e_temp}{Catch}

\item{b_temp}{Biomass}

\item{m_season}{natural mortality}

\item{f_sel}{Fisheries selectivity}

\item{n_temp}{Numbers at age}

\item{w_catch}{weight at age}

\item{method}{'Pope' or 'Hybrid' method?}
}
\value{
The fishing mortality value
}
\description{
Calculate the fishing mortality based on catch
}
