% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{get_likelihoods}
\alias{get_likelihoods}
\title{Get likelihood values for a model object the begin with 'ans'}
\usage{
get_likelihoods(report)
}
\arguments{
\item{report}{The object returned by the \code{report()} function of the returned object of the minimizer.}
}
\value{
A list of the likelihoods whose names begin with 'ans'. Add these in pacifichakemse.cpp at the end as a REPORT
}
\description{
Get likelihood values for a model object the begin with 'ans'
}
\examples{
\dontrun{
obj <- MakeADFun(d, p, DLL = "pacifichakemse", silent = FALSE)
report <- obj$report()
get_likelihoods(report)
}
}
