% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ref_point.R
\name{get_ref_point}
\alias{get_ref_point}
\title{Calculate the catch and F40 value to apply to the next year in the MSE loop. TACs will be applied
to the catch value.}
\usage{
get_ref_point(
  pars,
  df,
  yr = NULL,
  ssb_y,
  f_in = NA,
  n_end,
  tac = NULL,
  v_real = NA,
  space = 2,
  catch_floor = NULL,
  ...
)
}
\arguments{
\item{pars}{Estimated parameters}

\item{df}{Data frame of non-estimated parameters}

\item{yr}{The year to calculate the reference points for}

\item{ssb_y}{Spawning biomass}

\item{f_in}{Fishing mortality}

\item{n_end}{Numbers at age}

\item{tac}{TAC values to apply to the catch before returning the function.}

\item{v_real}{Vulnerable biomass in OM}

\item{space}{The space (area) to use the selectivity values for. Defaults to USA (2)}

\item{catch_floor}{The lowest catch to be allowed}

\item{...}{Absorb arguments meant for other functions}
}
\value{
A list of two, catch and F40 for the next year
}
\description{
Calculate the catch and F40 value to apply to the next year in the MSE loop. TACs will be applied
to the catch value.
}
