% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSelec.R
\name{get_select}
\alias{get_select}
\title{Get a vector of selectivities of the same length as the \code{age} vector
based on a range asked for}
\usage{
get_select(ages = NULL, p_sel = NULL, s_min = NULL, s_max = NULL)
}
\arguments{
\item{p_sel}{Log selectivity values from age \code{s_min} to \code{s_max}}

\item{s_min}{Minimum age}

\item{s_max}{Maximum age}

\item{age}{Vector of ages}
}
\value{
A vector of selectivity values of the same length as the \code{age} vector
}
\description{
Get a vector of selectivities of the same length as the \code{age} vector
based on a range asked for
}
