% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{get_yrs_mse_list}
\alias{get_yrs_mse_list}
\title{Get a vector of the years from a list of MSEs}
\usage{
get_yrs_mse_list(lst)
}
\arguments{
\item{lst}{A list of MSE results, constructed from code such as:
ls_plots <- map(fls,~{readRDS(.x)}) where fls is a list of RDS filenames
which have been written by \code{\link[=run_mses]{run_mses()}}. See \code{\link[=setup_mse_plot_objects]{setup_mse_plot_objects()}}
for a closer look at how this list is constructed}
}
\value{
A vector of the years from a list of MSEs
}
\description{
Get a vector of the years from a list of MSEs
}
