% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getF.R
\name{getF}
\alias{getF}
\title{Calculate the fishing mortality based on catch}
\usage{
getF(Etemp, Btemp, Mseason, Fsel, Ntemp, w_catch = w_catch, method = "Pope")
}
\arguments{
\item{method}{}
}
\value{

}
\description{
Calculate the fishing mortality based on catch
}
