% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getParameters.R
\name{getParameters}
\alias{getParameters}
\title{Load parameters for \link{TMB} model}
\usage{
getParameters(trueparms = TRUE, df)
}
\arguments{
\item{trueparms}{use the correct parameters for initial distribution?}

\item{df}{data frame of OM parameters and life history}
}
\value{
A list of parameters
}
\description{
Load parameters for \link{TMB} model
}
