% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getParameters_csv.R
\name{getParameters_csv}
\alias{getParameters_csv}
\title{Export EM parameters for \code{\link[TMB:MakeADFun]{TMB::MakeADFun()}} based on csv files with parameters}
\usage{
getParameters_csv(trueparms = TRUE)
}
\arguments{
\item{trueparms}{Logical. TRUE if random parameter initialization}
}
\value{
A list of parameters
}
\description{
Export EM parameters for \code{\link[TMB:MakeADFun]{TMB::MakeADFun()}} based on csv files with parameters
}
