% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getRefpoint.R
\name{getRefpoint}
\alias{getRefpoint}
\title{Get reference points (TODO: improve docs on this function)}
\usage{
getRefpoint(par_fixed, df, ssb_y, f_in = NA, n_end, tac = 1, v_real = NA)
}
\arguments{
\item{par_fixed}{estimated parameters}

\item{df}{data frame of non-estimated parameters}

\item{ssb_y}{Spawning biomass}

\item{f_in}{Fishing mortality}

\item{n_end}{Numbers at age}

\item{tac}{which tac function to use?}

\item{v_real}{Vulnerable biomass in OM}
}
\value{
A list of reference points
}
\description{
Get reference points (TODO: improve docs on this function)
}
