% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSelec.R
\name{getSelec}
\alias{getSelec}
\title{Title}
\usage{
getSelec(age, psel, Smin, Smax)
}
\arguments{
\item{age}{vector of ages}

\item{psel}{selectivity parameters}

\item{Smin}{Minimum caught age}

\item{Smax}{maximum age to model seperately}
}
\value{

}
\description{
Title
}
\examples{

}
