% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_agebased_model.R
\name{init_agebased_model}
\alias{init_agebased_model}
\title{Initialize values used in the age based model}
\usage{
init_agebased_model(om = NULL)
}
\arguments{
\item{om}{A \link{list} as returned by \code{\link[=load_data_om]{load_data_om()}}}
}
\value{
A \link{list} with the same structure as \code{om}, but with a new element,
\code{n_init} added, and some values changed for the initialization of the model
}
\description{
Initialize values used in the age based model
}
