% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init-movement-mat.R
\name{init_movement_mat}
\alias{init_movement_mat}
\title{Initialize the movement model matrix. An alternative function should be
written if changes are required to the initialization assumptions}
\usage{
init_movement_mat(
  n_space = NULL,
  space_names = NULL,
  n_season = NULL,
  season_names = NULL,
  m_yr = NULL,
  n_yr = NULL,
  yrs = NULL,
  move_max = NULL,
  move_slope = NULL,
  move_fifty = NULL,
  move_south = NULL,
  move_out = NULL,
  move_init = NULL,
  ages_no_move = NULL,
  ages = NULL,
  age_names = NULL,
  f_space = NULL
)
}
\arguments{
\item{n_space}{See \code{\link[=load_data_om]{load_data_om()}}}

\item{space_names}{See \code{\link[=load_data_om]{load_data_om()}}}

\item{n_season}{See \code{\link[=load_data_om]{load_data_om()}}}

\item{season_names}{See \code{\link[=load_data_om]{load_data_om()}}}

\item{m_yr}{The last non-future year. Used to de-populate future values
if \code{populate_future} is FALSE}

\item{n_yr}{The number of years in the array dimension}

\item{yrs}{A vector of names for the years. Length must equal \code{n_yr}}

\item{move_max}{A vector of the maximum movement rate, one for each
of \code{n_seasons}}

\item{move_slope}{See \code{\link[=load_data_om]{load_data_om()}}}

\item{move_fifty}{Age at 50 percent movement rate}

\item{move_south}{See \code{\link[=load_data_om]{load_data_om()}}}

\item{move_out}{See \code{\link[=load_data_om]{load_data_om()}}}

\item{move_init}{See \code{\link[=load_data_om]{load_data_om()}}}

\item{ages_no_move}{See \code{\link[=load_data_om]{load_data_om()}}}

\item{ages}{See \code{\link[=load_data_om]{load_data_om()}}}

\item{age_names}{See \code{\link[=load_data_om]{load_data_om()}}}

\item{f_space}{See \code{\link[=load_data_om]{load_data_om()}}}
}
\value{
A list of 3 elements: The \code{move_mat} matrix for movement,
the \code{move_init} vector of length \code{n_space} and the \code{f_space} vector
of length \code{n_space}
}
\description{
Initialize the movement model matrix. An alternative function should be
written if changes are required to the initialization assumptions
}
