% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data.R
\name{load_data}
\alias{load_data}
\title{Load hake data for operating model}
\usage{
load_data()
}
\value{
A list of Parameters, Input parameters, Survey, Catch, and others
}
\description{
Load hake data for operating model
}
