% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data_seasons.R
\name{load_data_seasons}
\alias{load_data_seasons}
\title{Load the hake data for seasons (TODO: Improve docs on this function)}
\usage{
load_data_seasons(
  nseason = 4,
  nspace = 2,
  syr = 1966,
  myr = 2018,
  ages = 0:20,
  sel_change_yr = 1991,
  movemaxinit = 0.35,
  movefiftyinit = 6,
  nsurveys = 2,
  logSDR = 1.4,
  bfuture = 0.5,
  moveout = 0.85,
  movesouth = 0.05,
  moveinit = NA,
  moveslope = 0.9,
  selectivity_change = 0,
  yr_future = 0,
  sel_hist = TRUE
)
}
\arguments{
\item{nseason}{Number of seasons}

\item{nspace}{Number of spatial areas}

\item{syr}{First year of historical simulations}

\item{myr}{Last year of historical simulations}

\item{ages}{A vector of ages}

\item{sel_change_yr}{A year in which a selectivity change took place}

\item{movemaxinit}{Maximum movement rate}

\item{movefiftyinit}{Age at 50 percent maximum movement rate}

\item{logSDR}{Recruitment deviations}

\item{bfuture}{Bias adjustment in the future}

\item{moveout}{Fraction of individuals that travel south in the last season}

\item{movesouth}{Fraction of individuals that move south during the year}

\item{moveinit}{Initial distribution of fish}

\item{moveslope}{Slope of the movement function}

\item{selectivity_change}{Should selectivity change?}

\item{yr_future}{How many years into the future should there be stochastic values}

\item{sel_hist}{Use historical selectivity?}

\item{nsurvey}{Survey frequency}
}
\value{
A list of Parameters, Input parameters, Survey, Catch, and others
}
\description{
Load the hake data for seasons (TODO: Improve docs on this function)
}
\examples{
\dontrun{
df <- load_data_seasons(nseason = 12, nspace = 2)
}
}
