% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data_ss.R
\name{load_data_ss}
\alias{load_data_ss}
\title{Load the hake data}
\usage{
load_data_ss(mod, sum_zero = 0)
}
\arguments{
\item{mod}{SS3 file}

\item{sum_zero}{flag on whether 'main' recruitment deviations should sum to zero}
}
\value{
A list of parameters, input parameters, catch, selectivity
}
\description{
Load the hake data
}
