% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-load-data-table-from-web.R
\name{load_data_table_from_web}
\alias{load_data_table_from_web}
\title{Read in a data table from a CSV data file at a URL on the web}
\usage{
load_data_table_from_web(url)
}
\arguments{
\item{url}{The full URL of the data file including the name of the file
and the .csv extension}
}
\value{
The data table as a \link[tibble:tibble]{tibble::tibble} or \code{NULL} if there was a
problem with the URL or loading from it
}
\description{
Read in a data table from a CSV data file at a URL on the web
}
