% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_mse_plot_data.R
\name{load_mse_plot_data}
\alias{load_mse_plot_data}
\title{Load an mse plot object RDS file which holds the output of \code{\link[=setup_mse_plot_objects]{setup_mse_plot_objects()}}}
\usage{
load_mse_plot_data(
  scenario = NULL,
  overwrite_rds = FALSE,
  main_results_dir = "results",
  quants = c(0.05, 0.5, 0.95),
  ...
)
}
\arguments{
\item{scenario}{The name of the scenario. This is the name of a subdirectory found in
the  \code{main_results_dir} directory of the project.}

\item{overwrite_rds}{Logical. If \code{TRUE} and the file exists, it will be overwritten. If \code{FALSE},
the file will be created if it does not exist and left as-is if it exists.}

\item{main_results_dir}{The name of the results directory which holds the mse output subdirectories
in the project.}

\item{quants}{Quantile values to use}

\item{...}{Arguments to send to \code{\link[=hake_objectives]{hake_objectives()}}}
}
\value{
The object which was returned by \code{\link[=setup_mse_plot_objects]{setup_mse_plot_objects()}}
}
\description{
Load an mse plot object RDS file which holds the output of \code{\link[=setup_mse_plot_objects]{setup_mse_plot_objects()}}
}
\details{
If the file exists and \code{overwrite_rds} is \code{FALSE}, the file is read in and the
content object is returned. If the file does not exist, it is created prior to the results
being returned.
}
