% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{load_parameters}
\alias{load_parameters}
\title{Load a list of parameter estimates from the SS model output}
\usage{
load_parameters(ss_model)
}
\arguments{
\item{ss_model}{SS3 model output as created by \code{\link[=create_rds_file]{create_rds_file()}}}
}
\value{
A named \link{list} of parameter estimates
}
\description{
Load a list of parameter estimates from the SS model output
}
