% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{load_sel_parameters}
\alias{load_sel_parameters}
\title{Load a selectivity parameter estimates from the SS model output}
\usage{
load_sel_parameters(ss_model = NULL)
}
\arguments{
\item{ss_model}{SS3 model output as created by \code{\link[=create_rds_file]{create_rds_file()}}}
}
\value{
A \link{data.frame} of estimates, age, and source (fishery and survey)
}
\description{
Load a selectivity parameter estimates from the SS model output
}
