% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load-ss-extract-age-comps.R
\name{load_ss_extract_age_comps}
\alias{load_ss_extract_age_comps}
\title{Extract the age comp estimates from the SS assessment model output into
a format required for input into the TMB assessment model}
\usage{
load_ss_extract_age_comps(
  ss_model = NULL,
  age_comps_fleet = 1,
  start_yr = NULL,
  end_yr = NULL,
  age_comps_fill = -1,
  yr_col = "yr",
  age_comps_fleet_col = "fleet",
  age_col = "age",
  value_col = "50\%",
  ...
)
}
\arguments{
\item{ss_model}{SS model input/output as read in by \code{\link[=load_ss_model_data]{load_ss_model_data()}}}

\item{age_comps_fleet}{1 for fishery, 2 for survey}

\item{start_yr}{If not \code{NULL}, remove all years prior to this}

\item{end_yr}{If not \code{NULL}, remove all years after this}

\item{age_comps_fill}{Value to replace NAs in the table, it can also be
\code{NA}, but not \code{NULL}}

\item{yr_col}{The name of the column in \code{ss_model$extra_mcmc$age_comps}
that contains the year}

\item{age_comps_fleet_col}{The name of the column in
\code{ss_model$extra_mcmc$age_comps} that contains the fleet code}

\item{...}{Arguments absorbed which are meant for other functions}
}
\value{
The matrix representing all years from \code{start_yr} to \code{end_yr} with
data included for years which are found in the \code{ss_model} data. Years
not in the \code{ss_model} data will be filled with the value of
\code{age_comps_fill}
}
\description{
Extract the age comp estimates from the SS assessment model output into
a format required for input into the TMB assessment model
}
\details{
Proportions at age for each year are calculated and returned
}
