% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ss-model-loading.R
\name{load_ss_model_from_rds}
\alias{load_ss_model_from_rds}
\title{Load models from files created using \code{\link[=create_rds_file]{create_rds_file()}}}
\usage{
load_ss_model_from_rds(model_dir = NULL, data_csv_dir = NULL, ...)
}
\arguments{
\item{model_dir}{A model directory name}

\item{data_csv_dir}{A directory containing the csv data for the assessment}

\item{...}{Arguments to be passed to \code{\link[=create_rds_file]{create_rds_file()}}}
}
\value{
A \link{list} of model inputs/outputs for the requested model
}
\description{
Load models from files created using \code{\link[=create_rds_file]{create_rds_file()}}
}
\details{
Load an SS3 model from an RDS file and return as a \link{list}
}
