% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load-ss-parameters.R
\name{load_ss_parameters}
\alias{load_ss_parameters}
\title{Load a list of parameter estimates from the SS model output}
\usage{
load_ss_parameters(ss_model = NULL)
}
\arguments{
\item{ss_model}{SS3 model output as created by \code{\link[hake:create_rds_file]{hake::create_rds_file()}}}
}
\value{
A named list of parameter estimates
}
\description{
Load a list of parameter estimates from the SS model output
}
