% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load-ss-sel-parameters.R
\name{load_ss_sel_parameters}
\alias{load_ss_sel_parameters}
\title{Load a selectivity parameter estimates from the SS model output}
\usage{
load_ss_sel_parameters(ss_model = NULL, ...)
}
\arguments{
\item{ss_model}{SS3 model output as created by \code{\link[=create_rds_file]{create_rds_file()}}}

\item{...}{Arguments absorbed which are meant for other functions}
}
\value{
A data frame of estimates (value column), age, and source
(fishery and survey are source 1 and 2 respectively)
}
\description{
Load a selectivity parameter estimates from the SS model output
}
