% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{modify_wage_df}
\alias{modify_wage_df}
\title{Modify the \code{yr} row in the \code{wage} \link{data.frame}, copying data from the \code{yr_copy} row}
\usage{
modify_wage_df(wage = NULL, yr = NULL, yr_copy = NULL)
}
\arguments{
\item{wage}{A weight-at-age \link{data.frame} as created by \code{\link[=load_ss_model_data]{load_ss_model_data()}}}

\item{yr}{The year to modify the row data for in the \link{data.frame}}

\item{yr_copy}{The year to use as source data row for the copy. If \code{NULL}, the first row will be used}
}
\value{
The same \code{wage} \link{data.frame} with the year row modified
}
\description{
Modify the \code{yr} row in the \code{wage} \link{data.frame}, copying data from the \code{yr_copy} row
}
