% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PacifichakeMSE-package.R
\docType{package}
\name{PacifichakeMSE-package}
\alias{PacifichakeMSE-package}
\title{Pacific hake MSE package}
\description{
\tabular{ll}{ Package: \tab PacifichakeMSE\cr Type: \tab Package\cr Version: \tab
1.0\cr Date: \tab 2020-02-26\cr License: \tab GPL-3\cr LazyLoad: \tab
yes\cr URL: \tab \url{https://github.com/PacifichakeMSE/}\cr }
}
\examples{

\dontrun{
# Run the MSE
Fill in stuff

# Run the hake stock assessment
runhakeassessment(123)
}

}
\references{
PacifichakeMSE on GitHub: \url{https://github.com/PacifichakeMSE}
}
\author{
Nis S Jacobsen, Ian G. Taylor, Kristin Marshall, Aaron Berger 
Package maintainer: Nis Jacobsen <nissandjac@gmail.com>
}
