% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PacifichakeMSE.R
\docType{package}
\name{PacifichakeMSE}
\alias{PacifichakeMSE}
\title{Pacific hake MSE package}
\examples{
\dontrun{
# Run the MSE
Fill in stuff

# Run the hake stock assessment
runhakeassessment(123)
}

}
\references{
PacifichakeMSE on GitHub: \url{https://github.com/PacifichakeMSE}
}
\author{
Nis S Jacobsen, Ian G. Taylor, Kristin Marshall, Aaron Berger
Package maintainer: Nis Jacobsen \href{mailto:nissandjac@gmail.com}{nissandjac@gmail.com}
}
