% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_bars_country_season.R
\name{plot_bars_country_season}
\alias{plot_bars_country_season}
\title{Plot faceted barplots by country, season, and scenarios}
\usage{
plot_bars_country_season(ps = NULL)
}
\arguments{
\item{ps}{A plot setup object as output by \code{\link[=setup_mse_plot_objects]{setup_mse_plot_objects()}}}
}
\value{
A \code{\link[ggplot2:ggplot]{ggplot2::ggplot()}} object
}
\description{
Plot faceted barplots by country, season, and scenarios
}
