% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_catch.R
\name{plot_catch}
\alias{plot_catch}
\title{Plot Catch by scenario and country}
\usage{
plot_catch(ps)
}
\arguments{
\item{ps}{A plot setup object as output by \code{\link[=setup_mse_plot_objects]{setup_mse_plot_objects()}}}
}
\value{
A \code{\link[ggplot2:ggplot]{ggplot2::ggplot()}} object
}
\description{
Plot Catch by scenario and country
}
