% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-em-timeseries.R, R/plot_em_timeseries.R
\name{plot_em_timeseries}
\alias{plot_em_timeseries}
\title{Plot Estimation model time series data from an MSE or OM simulation run}
\usage{
plot_em_timeseries(
  ps,
  type = "ssb",
  ci = c(0.05, 0.95),
  ci_lines = TRUE,
  yr_lim = c(NA_real_, NA_real_),
  rev_scenarios = TRUE,
  legend_position = c(0.15, 0.92),
  free_y_scale = FALSE,
  ...
)

plot_em_timeseries(
  ps,
  type = "ssb",
  ci = c(0.05, 0.95),
  ci_lines = TRUE,
  yr_lim = c(NA_real_, NA_real_),
  rev_scenarios = TRUE,
  legend_position = c(0.15, 0.92),
  free_y_scale = FALSE,
  ...
)
}
\arguments{
\item{ps}{A plot setup object as output by \code{\link[=setup_mse_plot_objects]{setup_mse_plot_objects()}}.}

\item{type}{One of 'ssb', 'ssb_ssb0', catch', 'aas', 'aac','aap'}

\item{ci}{A vector of length two of the lower and upper credible interval values.}

\item{ci_lines}{If TRUE, use dashed lines for the CI envelope. If FALSE, use a shaded ribbon.}

\item{yr_lim}{A vector of 2 for minimum and maximum yrs to show on the plot. If either are NA,
the limits of the data are used.}

\item{rev_scenarios}{If TRUE, reverse the order of the scenarios in the legend.}

\item{legend_position}{A vector of two - X and Y position for the legend. See \code{\link[ggplot2:theme]{ggplot2::theme()}}}

\item{free_y_scale}{See the \code{scales} argument of \code{\link[ggplot2:facet_wrap]{ggplot2::facet_wrap()}}}

\item{...}{Extra arguments to pass to \code{\link[=color_facet_backgrounds]{color_facet_backgrounds()}}}

\item{by_country}{If TRUE, the plot will be N panels of plots comparing the scenarios by country.
N is the number of scenarios. If FALSE the plot will be one panel comparing the scenarios.}

\item{ylim}{A vector of 2 for limits of the y-axis. If either are NA,
the limits of the data are used.}

\item{show_ssb0}{If TRUE, show the initial biomass, SSB0 line.
Only used if \code{type} is 'ssb' or 'ssb_ssb0'.}

\item{show_40_10}{If TRUE, show the 0.4 initial biomass and 0.1 initial biomass lines.
Only used if \code{type} is 'ssb' or 'ssb_ssb0'.}

\item{ssb_line_txt_cex}{Size multiplier for SSB reference point line labels.}

\item{ssb_line_col}{Color for SSB reference point line labels.}

\item{ssb_line_type}{Line type for SSB reference point line labels.}
}
\value{
A \code{\link[ggplot2:ggplot]{ggplot2::ggplot()}} object

A \code{\link[ggplot2:ggplot]{ggplot2::ggplot()}} object
}
\description{
Plot Estimation model time series data from an MSE or OM simulation run

Plot the mean of the year medians SSB for all Estimation models
}
