% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_om.R
\name{plot_selectivity_scenarios}
\alias{plot_selectivity_scenarios}
\title{Plot a 3-pane selectivity plot by country for three scenarios:
Base, Low US, and last year selectivity}
\usage{
plot_selectivity_scenarios(om_0, om_1, om_2, type = "new")
}
\arguments{
\item{om_0}{Operating model output from \code{\link[=run_om]{run_om()}}}

\item{om_1}{Operating model output from \code{\link[=run_om]{run_om()}}}

\item{om_2}{Operating model output from \code{\link[=run_om]{run_om()}}}

\item{type}{"new or "old"}
}
\value{
A \code{\link[ggplot2:ggplot]{ggplot2::ggplot()}} object
}
\description{
Plot a 3-pane selectivity plot by country for three scenarios:
Base, Low US, and last year selectivity
}
