% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_ssb.R
\name{plot_ssb}
\alias{plot_ssb}
\title{Plot Spawning Stock Biomass (SSB) by scenario all in one panel}
\usage{
plot_ssb(
  ps = NULL,
  time = "beg",
  ci = c(0.05, 0.95),
  yr_lim = c(NA_real_, NA_real_)
)
}
\arguments{
\item{ps}{A plot setup object as output by \code{\link[=setup_mse_plot_objects]{setup_mse_plot_objects()}}.}

\item{time}{Either 'beg' for beginning of the year SSB or 'mid' for mid-year SSB.}

\item{ci}{A vector of length two of the lower and upper credible interval values.}

\item{yr_lim}{A vector of 2 for minimum and maximum yrs to show on the plot. If either are NA,
the limits of the data are used}
}
\value{
A \code{\link[ggplot2:ggplot]{ggplot2::ggplot()}} object
}
\description{
Plot Spawning Stock Biomass (SSB) by scenario all in one panel
}
