% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_ssb_ssb0.R
\name{plot_ssb_ssb0}
\alias{plot_ssb_ssb0}
\title{Plot total SSB/SSB0 by scenario where SSB is the Spawning Stock Biomass from the
Estimation model and SSB0 is the initial SSB from teh Operating model}
\usage{
plot_ssb_ssb0(ps = NULL)
}
\arguments{
\item{ps}{A plot setup object as output by \code{\link[=setup_mse_plot_objects]{setup_mse_plot_objects()}}}
}
\value{
A \code{\link[ggplot2:ggplot]{ggplot2::ggplot()}} object
}
\description{
Plot total SSB/SSB0 by scenario where SSB is the Spawning Stock Biomass from the
Estimation model and SSB0 is the initial SSB from teh Operating model
}
