% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_standard_error.R
\name{plot_standard_error}
\alias{plot_standard_error}
\title{Plot Standard Error of the Spawning Stock Biomass between the Operating model and
the Estimation model}
\usage{
plot_standard_error(
  ps = NULL,
  ci = c(0.05, 0.95),
  yr_lim = c(NA_real_, NA_real_)
)
}
\arguments{
\item{ps}{A plot setup object as output by \code{\link[=setup_mse_plot_objects]{setup_mse_plot_objects()}}}

\item{ci}{A vector of length two of the lower and upper credible interval values.
These values will be passed to \code{\link[stats:quantile]{stats::quantile()}}}

\item{yr_lim}{A vector of 2 for minimum and maximum yrs to show on the plot. If either are NA,
the limits of the data are used}
}
\value{
A \code{\link[ggplot2:ggplot]{ggplot2::ggplot()}} object
}
\description{
Plot Standard Error of the Spawning Stock Biomass between the Operating model and
the Estimation model
}
