% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_tac_estimated_vs_realized.R
\name{plot_tac_est_vs_real}
\alias{plot_tac_est_vs_real}
\title{Plot TAC vs SSB as a simple line}
\usage{
plot_tac_est_vs_real(df)
}
\arguments{
\item{df}{A \link{data.frame} as output by \code{\link[=calc_tac_est_vs_real]{calc_tac_est_vs_real()}}}
}
\value{
A \code{\link[ggplot2:ggplot]{ggplot2::ggplot()}} object
}
\description{
Plot TAC vs SSB as a simple line
}
