% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_tac_estimated_vs_realized.R
\name{plot_tac_vs_hcr}
\alias{plot_tac_vs_hcr}
\title{Plot estimated vs realized TAC (called Catch in plot) by Harvest Control Rule}
\usage{
plot_tac_vs_hcr(ps = NULL)
}
\arguments{
\item{ps}{A plot setup object as output by \code{\link[=setup_mse_plot_objects]{setup_mse_plot_objects()}}}
}
\value{
A \code{\link[ggplot2:ggplot]{ggplot2::ggplot()}} object
}
\description{
Plot estimated vs realized TAC (called Catch in plot) by Harvest Control Rule
}
