% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_timeseries.R
\name{plot_timeseries}
\alias{plot_timeseries}
\title{Plot time series data from an MSE or OM simulation run}
\usage{
plot_timeseries(
  ps = NULL,
  type = "ssb",
  time = "mid",
  ci = c(0.05, 0.95),
  by_country = FALSE,
  yr_lim = c(NA_real_, NA_real_),
  ci_lines = TRUE,
  show_ssb0 = TRUE,
  show_40_10 = TRUE,
  rev_scenarios = FALSE,
  legend_position = c(0.08, 0.15),
  ...
)
}
\arguments{
\item{ps}{A plot setup object as output by \code{\link[=setup_mse_plot_objects]{setup_mse_plot_objects()}}.}

\item{type}{One of 'ssb', 'ssb_ssb0', catch', 'aas', 'aac','aap'}

\item{time}{Either 'beg' for beginning of the year SSB or 'mid' for mid-year SSB.
Only used if \code{type} is 'ssb' or 'ssb_ssb0'.}

\item{ci}{A vector of length two of the lower and upper credible interval values.}

\item{by_country}{If TRUE, the plot will be N panels of plots comparing the scenarios by country.
N is the number of scenarios. If FALSE the plot will be one panel comparing the scenarios.}

\item{yr_lim}{A vector of 2 for minimum and maximum yrs to show on the plot. If either are NA,
the limits of the data are used.}

\item{ci_lines}{If TRUE, use dashed lines for the CI envelope. If FALSE, use a shaded ribbon.}

\item{show_ssb0}{If TRUE, show the initial biomass, SSB0 line.
Only used if \code{type} is 'ssb' or 'ssb_ssb0'.}

\item{show_40_10}{If TRUE, show the 0.4 initial biomass and 0.1 initial biomass lines.
Only used if \code{type} is 'ssb' or 'ssb_ssb0'.}

\item{rev_scenarios}{If TRUE, reverse the order of the scenarios in the legend.}

\item{legend_position}{A vector of two - X and Y position for the legend. See \code{\link[ggplot2:theme]{ggplot2::theme()}}}

\item{...}{Extra arguments to pass to \code{\link[=color_facet_backgrounds]{color_facet_backgrounds()}}}
}
\value{
A \code{\link[ggplot2:ggplot]{ggplot2::ggplot()}} object
}
\description{
Plot time series data from an MSE or OM simulation run
}
