% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_timeseries_runs.R
\name{plot_timeseries_runs}
\alias{plot_timeseries_runs}
\title{Plot individual run output against each other as a time series}
\usage{
plot_timeseries_runs(
  ps,
  type = "ssb",
  scen = 1,
  yr_lim = c(NA_real_, NA_real_),
  ylim = c(NA_real_, NA_real_),
  legend_position = c(0.26, 0.85),
  line_width = 1,
  line_type = "solid"
)
}
\arguments{
\item{ps}{A plot setup object as output by \code{\link[=setup_mse_plot_objects]{setup_mse_plot_objects()}}.}

\item{type}{One of 'ssb', 'ssb_ssb0', catch', 'aas', 'aac','aap'}

\item{scen}{The scenario number to use. This is an integer representing the names given by \code{names(ps$sim_data)}}

\item{yr_lim}{A vector of 2 for minimum and maximum yrs to show on the plot. If either are NA,
the limits of the data are used.}

\item{ylim}{A vector of 2 for limits of the y-axis. If either are NA,
the limits of the data are used.}

\item{legend_position}{A vector of two - X and Y position for the legend. See \code{\link[ggplot2:theme]{ggplot2::theme()}}}

\item{line_width}{Width of the lines. See \code{\link[ggplot2:geom_path]{ggplot2::geom_line()}}}

\item{line_type}{Type of the lines. See \code{\link[ggplot2:geom_path]{ggplot2::geom_line()}}}
}
\description{
Plot individual run output against each other as a time series
}
