% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_violins_country.R
\name{plot_violins_country}
\alias{plot_violins_country}
\title{Plot faceted violin plots by Country for scenarios}
\usage{
plot_violins_country(
  ps = NULL,
  pidx = NULL,
  yrs = NULL,
  facet_back_cols = c("red", "blue"),
  ...
)
}
\arguments{
\item{ps}{A plot setup object as output by \code{\link[=setup_mse_plot_objects]{setup_mse_plot_objects()}}}

\item{pidx}{The order in which to plot the scenarios. If \code{NULL}, they will be in the
order of the \link{data.frame}}

\item{yrs}{The years to include in the plot. If \code{NULL}, all available data will be
used}
}
\value{
A \code{\link[ggplot2:ggplot]{ggplot2::ggplot()}} object
}
\description{
Plot faceted violin plots by Country for scenarios
}
