% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotMSE_biasadjustment.R
\name{plotMSE_biasadjustment}
\alias{plotMSE_biasadjustment}
\title{Plot MSE Bias adjustment (TODO: Improve docs on this function)}
\usage{
plotMSE_biasadjustment(results, plotnames = NA, plotexp = FALSE)
}
\arguments{
\item{results}{Output from... (TODO)}

\item{plotnames}{Names for the plots}

\item{plotexp}{Logical. If TRUE, save the two plots into files separately,
If FALSE, plot them using \code{\link[gridExtra:grid.arrange]{gridExtra::grid.arrange()}}}
}
\value{
A list of two \link{ggplot2} plots
}
\description{
Plot MSE Bias adjustment (TODO: Improve docs on this function)
}
