% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotSelectivity.R
\name{plotSelectivity}
\alias{plotSelectivity}
\title{Plot selectivity for years}
\usage{
plotSelectivity(df.new, parameters)
}
\arguments{
\item{df.new}{parameters and life history values}

\item{parameters}{estimates parameters}
}
\value{
Nothing
}
\description{
Plot selectivity for years
}
