% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotUncertainty.R
\name{plotUncertainty}
\alias{plotUncertainty}
\title{Plot uncertainty for a given parameter}
\usage{
plotUncertainty(est, data)
}
\arguments{
\item{est}{Parameter name}

\item{data}{A \link{data.frame} of uncertainty data}
}
\value{
Nothing
}
\description{
Plot uncertainty for a given parameter
}
