% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotValues.R
\name{plotValues}
\alias{plotValues}
\title{Plot the estimated or derived value from a TMB model}
\usage{
plotValues(est, data, name)
}
\arguments{
\item{est}{estimated values}

\item{data}{available data}

\item{name}{name to print in data frame}
}
\value{

}
\description{
Plot the estimated or derived value from a TMB model
}
