% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotViolin.R
\name{plotViolin}
\alias{plotViolin}
\title{Plot the estimated or derived value from a \link{TMB} model using a Violin-type plot}
\usage{
plotViolin(data, cols)
}
\arguments{
\item{data}{name of data set to be loaded}

\item{cols}{colors of the individual violins}
}
\value{
Nothing
}
\description{
Plot the estimated or derived value from a \link{TMB} model using a Violin-type plot
}
\examples{
\dontrun{
colors = c('black','green')
plotViolin(folder, colors) # plots black and green data
}
}
