% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-pr.R
\name{pr}
\alias{pr}
\title{Wrapper for the \code{print()} function}
\usage{
pr(d, num_rows = 100, ...)
}
\arguments{
\item{d}{The data frame}

\item{num_rows}{The number of rows to show on screen}

\item{...}{Arguments passed to \code{print()}}
}
\value{
Nothing, prints a data frame to screen
}
\description{
Wrapper for the \code{print()} function
}
