% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{round_3d_array}
\alias{round_3d_array}
\alias{round_4d_array}
\alias{round_5d_array}
\title{Round all numeric values found in a multidimensional \link{array}
to a specified number of decimal points}
\usage{
round_3d_array(arr, digits = 2)

round_4d_array(arr, digits = 2)

round_5d_array(arr, digits = 2)
}
\arguments{
\item{arr}{The array}

\item{digits}{The number of decimal points to round all numeric values to}
}
\value{
An array identical to the input \code{arr} but with all numerical values rounded
}
\description{
Round all numeric values found in a multidimensional \link{array}
to a specified number of decimal points
}
