% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{round_list}
\alias{round_list}
\title{Round all values in an arbitrarily complex \link{list}}
\usage{
round_list(lst, digits = 2)
}
\arguments{
\item{lst}{A \link{list} of arbitrary complexity}

\item{digits}{The number of decimal points to round all numeric values to}
}
\value{
A \link{list} in the same format as \code{lst} but with all values rounded to
\code{digits} decimal points
}
\description{
Round all values in an arbitrarily complex \link{list}
}
\details{
This is a recursive function and therefore can have an arbitrary
nesting of lists.
}
