% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_agebased_model_true_Catch.R
\name{run.agebased.true.catch}
\alias{run.agebased.true.catch}
\title{Run an agebased model}
\usage{
run.agebased.true.catch(df, seeds = 100)
}
\arguments{
\item{df}{data frame of parameters and life history values}

\item{seeds}{seed for survey error and recruitment deviations}
}
\value{

}
\description{
Run an agebased model
}
\examples{
run.agebased.true.catch(df) # runs the model

}
