% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_agebased_model_true_Catch.R
\name{run_agebased_true_catch}
\alias{run_agebased_true_catch}
\title{Run an agebased model}
\usage{
run_agebased_true_catch(df = NULL, seed = 100, ...)
}
\arguments{
\item{df}{data frame of parameters and life history values}

\item{seed}{seed for survey error and recruitment deviations}

\item{...}{Arguments passed to \code{\link[=run_year_loop_om]{run_year_loop_om()}}}
}
\value{
A list of model outputs (TODO)
}
\description{
Run an agebased model
}
\examples{
\dontrun{
run_agebased_true_catch(df)
}
}
