% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_mse_climate.R
\name{run_mse_climate}
\alias{run_mse_climate}
\title{Run the MSE for the Climate scenarios}
\usage{
run_mse_climate(
  ss_extdata_dir = NULL,
  nruns = 10,
  simyears = 30,
  fns = NULL,
  tacs = NULL,
  cincreases = NULL,
  mincreases = NULL,
  om_params_seed = 12345,
  results_root_dir = "results",
  results_dir = "climate",
  ...
)
}
\arguments{
\item{ss_extdata_dir}{A directory in the PacifichakeMSE package extdata directory
which contains SS3 model run output}

\item{nruns}{Then number of runs to do for each simulation}

\item{simyears}{The number of years to simulate into the future}

\item{fns}{A vector of file names for the scenarios (.rds files)}

\item{tacs}{A vector of TAC values to be passed to the \code{\link[=run_multiple_MSEs]{run_multiple_MSEs()}} function, in the same
order as the \code{fns} file names}

\item{cincreases}{A vector of values to be passed to the \code{\link[=run_multiple_MSEs]{run_multiple_MSEs()}} function, in the same
order as the \code{fns} file names}

\item{mincreases}{A vector of values to be passed to the \code{\link[=run_multiple_MSEs]{run_multiple_MSEs()}} function, in the same
order as the \code{fns} file names}

\item{om_params_seed}{A seed value to use when calling the \code{\link[=run.agebased.true.catch]{run.agebased.true.catch()}} function}

\item{results_root_dir}{The name of the results root directory (relative to \code{\link[here:here]{here::here()}})}

\item{results_dir}{The name of the results directory (relative to \link[here:here(results_root_dir)]{here::here(results_root_dir)})}

\item{...}{Arguments passed to \code{\link[=load_data_seasons]{load_data_seasons()}}}
}
\value{

}
\description{
Run the MSE for the Climate scenarios
}
\details{
Saves the output of the MSE runs to the files specified in \code{fns}
}
