% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_mse_hcr.R
\name{run_mse_hcr}
\alias{run_mse_hcr}
\title{Run the MSE for the Harvest Control Rule (HCR) scenarios}
\usage{
run_mse_hcr(
  ss_extdata_dir = NULL,
  nruns = 10,
  simyears = 30,
  fns = NULL,
  om_params_seed = 12345,
  results_root_dir = "results",
  results_dir = "hcr",
  ...
)
}
\arguments{
\item{ss_extdata_dir}{A directory in the PacifichakeMSE package extdata directory
which contains SS3 model run output}

\item{nruns}{Then number of runs to do for each simulation}

\item{simyears}{The number of years to simulate into the future}

\item{fns}{A vector of file names for the scenarios (.rds files)}

\item{om_params_seed}{A seed value to use when calling the \code{\link[=run.agebased.true.catch]{run.agebased.true.catch()}} function}

\item{results_root_dir}{The name of the results root directory (relative to \code{\link[here:here]{here::here()}})}

\item{results_dir}{The name of the results directory (relative to \link[here:here(results_root_dir)]{here::here(results_root_dir)})}

\item{...}{Arguments passed to \code{\link[=load_data_seasons]{load_data_seasons()}}}
}
\value{
Nothing
}
\description{
Run the MSE for the Harvest Control Rule (HCR) scenarios
}
\details{
Saves the output of the MSE runs to the files specified in \code{fns}
}
