% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_mses.R
\name{run_mses}
\alias{run_mses}
\title{Run the MSE for a number of runs for any scenarios}
\usage{
run_mses(
  ss_model_output_dir = NULL,
  n_runs = 10,
  n_sim_yrs = NULL,
  fns = NULL,
  plot_names = NULL,
  tacs = 1,
  c_increases = 0,
  m_increases = 0,
  sel_changes = 0,
  n_surveys = NULL,
  multiple_season_data = NULL,
  om_params_seed = 12345,
  results_root_dir = here("results"),
  results_dir = here("results", "default"),
  ...
)
}
\arguments{
\item{ss_model_output_dir}{A directory which contains SS3 model run output. Used to import
estimates from the SS3 model}

\item{n_runs}{Then number of runs to do for each simulation}

\item{n_sim_yrs}{The number of years to simulate into the future}

\item{fns}{A vector of file names for the scenarios (.rds files). .rds extension is optional}

\item{plot_names}{A vector of strings to use for the scenarios later when plotting. Must either be
\code{NULL} or the same length as \code{fns}}

\item{tacs}{A vector of TAC values to be passed to the \code{\link[=run_multiple_MSEs]{run_multiple_MSEs()}} function, in the same
order as the \code{fns} file names, or a single value}

\item{c_increases}{A vector of values to be passed to the \code{\link[=run_multiple_MSEs]{run_multiple_MSEs()}} function, in the same
order as the \code{fns} file names, or a single value}

\item{m_increases}{A vector of values to be passed to the \code{\link[=run_multiple_MSEs]{run_multiple_MSEs()}} function, in the same
order as the \code{fns} file names, or a single value}

\item{sel_changes}{A vector of values to be passed to the \code{\link[=run_multiple_MSEs]{run_multiple_MSEs()}} function, in the same
order as the \code{fns} file names, or a single value}

\item{n_surveys}{The number of surveys for each run. This must be a vector of the same length as \code{fns} or \code{NULL}.
If \code{NULL}, 2 will be used for every scenario}

\item{multiple_season_data}{A list of the same length as \code{fns}, with each element being a vector of
three items, \code{nseason}, \code{nspace}, and \code{bfuture}. If NULL, biasadjustment will not be incorporated}

\item{om_params_seed}{A seed value to use when calling the \code{\link[=run_om]{run_om()}} function}

\item{results_root_dir}{The results root directory}

\item{results_dir}{The results directory}

\item{...}{Arguments passed to \code{\link[=load_data_om]{load_data_om()}}}
}
\value{
Nothing
}
\description{
Run the MSE for a number of runs for any scenarios
}
\details{
Saves the output of the MSE runs to the files specified in \code{fns}, in the directories
specified in \code{results_root_dir} and \code{results_dir}
}
