% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_multiple_MSEs.R
\name{run_multiple_MSEs}
\alias{run_multiple_MSEs}
\title{Run/Iterate the Pacific hake MSE}
\usage{
run_multiple_MSEs(
  results_dir = NULL,
  file_name = NULL,
  df = NULL,
  ss_model = NULL,
  om_objs = NULL,
  random_seed = 12345,
  n_sim_yrs = NULL,
  tac = 1,
  c_increase = 0,
  m_increase = 0,
  sel_change = 0,
  f_sim = NULL,
  ...
)
}
\arguments{
\item{results_dir}{Directory in which the OM output will be stored}

\item{file_name}{The name of the file in which the MSE output was stored. This will be prepended with 'om_'
for the OM used in this MSE. The actual MSE results are stored by the function calling this one, \code{\link[=run_mses]{run_mses()}}}

\item{df}{Data frame of parameters as output by \code{\link[=load_data_om]{load_data_om()}}}

\item{ss_model}{SS3 model output as created by \code{\link[=create_rds_file]{create_rds_file()}}
and loaded by \code{\link[=load_ss_model_from_rds]{load_ss_model_from_rds()}}}

\item{om_objs}{Operating model objects as created by \code{\link[=setup_blank_om_objects]{setup_blank_om_objects()}}}

\item{random_seed}{Seed for running the OM if it needs to be run (if \code{sim_data}
is \code{NULL})}

\item{n_sim_yrs}{Number of years to simulate}

\item{tac}{Which harvest control rule should the model use}

\item{c_increase}{Increase in max movement}

\item{m_increase}{Decrease of spawners returning south}

\item{sel_change}{Time varying selectivity}

\item{f_sim}{Value of F to use for simulation years}

\item{...}{Absorb arguments intended for other functions}
}
\value{
A list of Catch, Catch.quota, SSB, SSB.mid, SSB.hes, Survey.om
F0, parms, N, converge, ams, amc, V
}
\description{
Run/Iterate the Pacific hake MSE
}
