% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_multiple_OMs.R
\name{run_multiple_OMs}
\alias{run_multiple_OMs}
\title{Run the Operating Models}
\usage{
run_multiple_OMs(
  n_sim_yrs = 30,
  seed = 12345,
  df,
  catch_in,
  c_increase = 0,
  m_increase = 0,
  sel_change = 0
)
}
\arguments{
\item{n_sim_yrs}{number of years to simulate}

\item{seed}{set the seed}

\item{df}{data frame of parameters}

\item{catch_in}{catches in the future}

\item{c_increase}{increase in movement due to climate change}

\item{m_increase}{increase in the number of returning spawners}

\item{sel_change}{should selectivity be time varying?}
}
\value{
A list of results
}
\description{
Run the Operating Models
}
\examples{
\dontrun{
run_multiple_OMs(df, 100) Catch is 100 kg in the future
}
}
