% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_om.R
\name{run_om}
\alias{run_om}
\title{Run the Operating Model, which is an age based model
conditioned on true catch}
\usage{
run_om(om = NULL, random_seed = NULL, n_sim_yrs = 0, ...)
}
\arguments{
\item{om}{list of parameters and life history values OM data. This list is
what is returned by \code{\link[=load_data_om]{load_data_om()}}}

\item{random_seed}{The random seed to use for this Operating Model}

\item{n_sim_yrs}{The number of future years to run}

\item{...}{Arguments passed to \code{\link[=run_year_loop_om]{run_year_loop_om()}}}
}
\value{
A list of model outputs
}
\description{
Run the Operating Model, which is an age based model
conditioned on true catch
}
