% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_oms.R
\name{run_oms}
\alias{run_oms}
\title{Run OMs for a number of runs}
\usage{
run_oms(
  ss_model = NULL,
  n_runs = NULL,
  yr_future = 0,
  fns = NULL,
  n_surveys = 2,
  b_futures = 0.5,
  sel_changes = 0,
  catch_in = NA_real_,
  hcr_apply = TRUE,
  random_recruitment = TRUE,
  attain = c(1, 1),
  plot_names = NULL,
  random_seed = NULL,
  results_root_dir = here("results"),
  results_dir = here("results", "default"),
  ...
)
}
\arguments{
\item{ss_model}{SS model data as returned from \code{\link[=load_ss_model_data]{load_ss_model_data()}}}

\item{n_runs}{The number of runs to run each scenario}

\item{yr_future}{Number of years to run the simulations out to}

\item{fns}{A vector of scenario names. See \code{\link[=run_mses]{run_mses()}}}

\item{n_surveys}{The number of surveys. See \code{\link[=run_mses]{run_mses()}}}

\item{b_futures}{Bias adjustment factors. See \code{\link[=run_mses]{run_mses()}}}

\item{sel_changes}{Selectivity changes. See \code{\link[=run_mses]{run_mses()}}}

\item{catch_in}{Catch into the future in kg}

\item{plot_names}{A vector of names to show on plots. See \code{\link[=run_mses]{run_mses()}}}

\item{random_seed}{The random seed to base run random seeds from. See \code{\link[=run_mses]{run_mses()}}}

\item{results_root_dir}{The results root directory}

\item{results_dir}{The results directory}

\item{...}{Arguments passed to \code{\link[=load_data_om]{load_data_om()}}}
}
\value{
Nothing, save OM output to a file
}
\description{
Run OMs for a number of runs
}
