% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runfuture_OM.R
\name{runfuture_OM}
\alias{runfuture_OM}
\title{Run simulatons of the operating model into the future}
\usage{
runfuture_OM(df, nruns = 100)
}
\arguments{
\item{df}{Data frame of}

\item{nruns}{number of iterations}
}
\value{
list of results
}
\description{
Run simulatons of the operating model into the future
}
\examples{

runfuture_OM(df, 10) #  makes 10 iterations
}
