% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{setup_blank_om_objects}
\alias{setup_blank_om_objects}
\title{Construct blank objects for Operating model outputs}
\usage{
setup_blank_om_objects(yrs, ages, nspace, nseason)
}
\arguments{
\item{yrs}{A vector of years}

\item{ages}{A vector of ages}

\item{nspace}{The number of spaces in the movement model}

\item{nseason}{The number of seasons in the movement model}
}
\value{
A list of the blank OM outputs
}
\description{
Construct blank objects for Operating model outputs
}
