% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotMSE.R
\name{setup_mse_plot_objects}
\alias{setup_mse_plot_objects}
\title{Setup MSE plot data frames and colors}
\usage{
setup_mse_plot_objects(
  results_dir = NULL,
  plotnames = NULL,
  plotexp = FALSE,
  pidx = NA
)
}
\arguments{
\item{results_dir}{A directory name where results from \code{\link[=run_mses]{run_mses()}} resides. Only files
in this directory that end in '.rds' and contain 'MSE' will be loaded}

\item{plotnames}{Names for the plots}

\item{plotexp}{See \code{\link[=fn_plot_MSE]{fn_plot_MSE()}}}

\item{pidx}{See \code{\link[=fn_plot_MSE]{fn_plot_MSE()}}}
}
\value{
A list of length 4: The items are (1) A \link{data.frame} containing the SSB/AAV/Catch
indicators, (2) A \link{data.frame} containing the country and season indicators, (3) A \link{data.frame}
containing the violin indicators (data in format for violin plots), (4) A \link{data.frame} of data
to be used to create violin plots, (5) A vector of colors, one for each file loaded (scenario),
(6) A vector of plot names, one for each scenario, and (7) sim_data, which is the output of
running \code{\link[=run.agebased.true.catch]{run.agebased.true.catch()}}
}
\description{
Setup MSE plot data frames and colors
}
\details{
Reads in the \code{.rds} files found in the \code{results_dir} directory and
creates two indicator \link{data.frame}s and colors for plotting
}
