% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarizeMSE.R
\name{summarizeMSE}
\alias{summarizeMSE}
\title{Title}
\usage{
summarizeMSE(ls.save, perc = c(0.1, 0.9))
}
\arguments{
\item{ls.save}{a saved list of MSE's}

\item{perc}{Percentiles to save}
}
\value{
Returns summaeized data
}
\description{
Title
}
\examples{
summarizeMSE(ls.save) # Get a summary of the MSE
}
