% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_perf_metrics.R
\name{table_perf_metrics}
\alias{table_perf_metrics}
\title{Create a \code{\link[knitr:kable]{knitr::kable()}} table object}
\usage{
table_perf_metrics(
  ps,
  ci = c(0.05, 0.5, 0.95),
  decimals = 0,
  scen = NULL,
  inc_mean = TRUE,
  format = "latex",
  ...
)
}
\arguments{
\item{ps}{A plot setup object as output by \code{\link[=setup_mse_plot_objects]{setup_mse_plot_objects()}}.}

\item{ci}{A vector of credible interval values. 0.5 must be included if the median is desired.}

\item{decimals}{The number of decimals to round values to}

\item{scen}{The scenario number to use. This is an integer representing the names given by \code{names(ps$sim_data)}}

\item{inc_mean}{If TRUE, include the mean in the table}

\item{format}{See \code{\link[knitr:kable]{knitr::kable()}}}

\item{...}{Arguments passed to \code{\link[knitr:kable]{knitr::kable()}}}
}
\value{
a \code{\link[knitr:kable]{knitr::kable()}} table object
}
\description{
Create a \code{\link[knitr:kable]{knitr::kable()}} table object
}
