% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_timeseries.R
\name{table_timeseries}
\alias{table_timeseries}
\title{Create a \code{\link[knitr:kable]{knitr::kable()}} table object}
\usage{
table_timeseries(
  ps = NULL,
  type = "ssb",
  time = "mid",
  ci = c(0.05, 0.5, 0.95),
  decimals = 0,
  by_country = FALSE,
  scen = NULL,
  yr_lim = c(NA_real_, NA_real_),
  inc_mean = TRUE,
  format = "latex",
  ...
)
}
\arguments{
\item{ps}{A plot setup object as output by \code{\link[=setup_mse_plot_objects]{setup_mse_plot_objects()}}.}

\item{type}{One of 'ssb', 'ssb_ssb0', catch', 'aas', 'aac'}

\item{time}{Either 'beg' for beginning of the year SSB or 'mid' for mid-year SSB.
Only used if \code{type} is 'ssb' or 'ssb_ssb0'.}

\item{ci}{A vector of credible interval values. 0.5 must be included if the median is desired.}

\item{decimals}{The number of decimals to round values to}

\item{by_country}{If TRUE, a table with country data side-by-side will be produced. If FALSE,
a table with aggregated output will be produced}

\item{scen}{The scenario number to use. This is an integer representing the names given by \code{names(ps$sim_data)}}

\item{yr_lim}{A vector of 2 for minimum and maximum yrs to include in the table. If either are NA,
the limits of the data are used.}

\item{inc_mean}{If TRUE, include the mean in the table}

\item{format}{See \code{\link[knitr:kable]{knitr::kable()}}}

\item{...}{Arguments passed to \code{\link[knitr:kable]{knitr::kable()}}}
}
\value{
a \code{\link[knitr:kable]{knitr::kable()}} table object
}
\description{
Create a \code{\link[knitr:kable]{knitr::kable()}} table object
}
