% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_om_data.R
\name{update_om_data}
\alias{update_om_data}
\title{Update the OM data for a new year}
\usage{
update_om_data(yr = NULL, yr_ind, yr_survey_sims, sim_data = NULL, df = NULL)
}
\arguments{
\item{yr}{The year to add}

\item{sim_data}{Operating model}

\item{df}{Input data and parameters
and loaded by \code{\link[=load_ss_model_from_rds]{load_ss_model_from_rds()}}}
}
\value{
A list of the data needed by \code{\link[TMB:MakeADFun]{TMB::MakeADFun()}}
}
\description{
Update the OM data for a new year
}
