% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{verify_argument}
\alias{verify_argument}
\title{Verify that the argument \code{arg} is valid in the context of the arguments given}
\usage{
verify_argument(arg = NULL, chk_class = NULL, chk_len = NULL, chk_is_in = NULL)
}
\arguments{
\item{arg}{The object to check}

\item{chk_class}{A character string of the name of the class to ensure \code{arg} is}

\item{chk_len}{A numeric value to ensure \code{arg} has the length of}

\item{chk_is_in}{A vector of values to ensure \code{arg} is in}
}
\value{
TRUE, invisibly or the function will throw an error if \code{arg} does not follow
the constraints given; FALSE is not returned
}
\description{
Verify that the argument \code{arg} is valid in the context of the arguments given
}
\examples{
verify_argument(23, "numeric", 1) # Succeeds
verify_argument(23, "numeric", 1, 1:20) # Fails
}
