% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{wage_add_yr}
\alias{wage_add_yr}
\title{Add a new row to the bottom of the \code{wage} \link{data.frame}, which is a copy of the
last row with the \code{Yr} value being one more than the last row's}
\usage{
wage_add_yr(wage = NULL)
}
\arguments{
\item{wage}{A weight-at-age \link{data.frame} as created by \code{\link[=load_data_ss]{load_data_ss()}}}
}
\value{
The same \code{wage} \link{data.frame} with a new row added to the bottom,
which is a copy of the last row with the \code{Yr} value being one more than
the last row's
}
\description{
Add a new row to the bottom of the \code{wage} \link{data.frame}, which is a copy of the
last row with the \code{Yr} value being one more than the last row's
}
