% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-calc-catch-seas-country.R
\name{calc_catch_seas_country}
\alias{calc_catch_seas_country}
\title{Calculate the catch proportions by country, year, and season}
\usage{
calc_catch_seas_country(
  data_tables_url = NULL,
  n_yrs = 10,
  weight = c(1, 1, 1, 1),
  ...
)
}
\arguments{
\item{n_yrs}{If \code{NULL}, a data frame containing all years of data with
proportions of catch by season will be returned. If this is a number, it is
the number of years at the end of the data to average the proportions by
season for. In this case, the data frame will be only 8 rows long, without
any year column}

\item{weight}{A vector of 4 for the weights given to each season}

\item{...}{Absorbs arguments meant for other functions}

\item{ss_model_data_csv_dir}{The directory in which the assessment csv
files containing catch by month are located. Can be a URL}
}
\value{
If \code{n_yrs} is \code{NULL}, a data frame containing all years of data with
proportions of catch by country and season. If \code{n_yrs} is a number, the
data frame returned will be only 8 rows long, 4 rows per country, without
any year column
}
\description{
Calculate the catch proportions by country, year, and season
}
