% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-calc-standard-error-ssb.R
\name{calc_standard_error_ssb}
\alias{calc_standard_error_ssb}
\title{Calculate the standard error on SSB between the Operating Model and the
Estimation Model for all runs in the list}
\usage{
calc_standard_error_ssb(
  em_output,
  om_output,
  col = "SE.SSB",
  quants = c(0.05, 0.25, 0.5, 0.75, 0.95),
  inc_mean = TRUE
)
}
\arguments{
\item{em_output}{A list of Estimation model output}

\item{om_output}{A list of Operating model outpu}

\item{col}{Column to calculate the quantiles on}

\item{quants}{A vector of quantiles to pass to \code{\link[stats:quantile]{stats::quantile()}}}

\item{inc_mean}{Logical. Include the mean (as column \code{avg})}
}
\value{
A \link{data.frame} of 3 columns: The run number, the standard error
of spawning stock biomass from EM in regards to the OM, and the year
}
\description{
Calculate the standard error on SSB between the Operating Model and the
Estimation Model for all runs in the list
}
