% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcMeanAge.R
\name{calcMeanAge}
\alias{calcMeanAge}
\title{Title}
\usage{
calcMeanAge(agemat, maxage)
}
\arguments{
\item{agemat}{matrix of ages}

\item{maxage}{maximum age to include in the calculation}
}
\value{

}
\description{
Title
}
\examples{
p <- calcMeanAge(matrix(1,1), 5)
Calculate mean age in the population
}
