% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcSE.R
\name{calcSE}
\alias{calcSE}
\title{Return the standard error of spawning stock biomass from EM in regards to the OM}
\usage{
calcSE(ls.MSE)
}
\description{
Return the standard error of spawning stock biomass from EM in regards to the OM
}
