% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcSE.R
\name{calcSE}
\alias{calcSE}
\title{Title}
\usage{
calcSE(ls.MSE)
}
\arguments{
\item{ls.MSE}{list of MSE results}
}
\value{

}
\description{
Title
}
\examples{
Return the standard error of spawning stock biomass from EM in regards to the OM
}
