% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Check_Identifiable_vs2.R
\name{Check_Identifiable_vs2}
\alias{Check_Identifiable_vs2}
\title{Check parameter convergence in TMB model}
\usage{
Check_Identifiable_vs2(obj)
}
\description{
Check parameter convergence in TMB model
}
