% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Check_Identifiable_vs2.R
\name{Check_Identifiable_vs2}
\alias{Check_Identifiable_vs2}
\title{Title}
\usage{
Check_Identifiable_vs2(obj, printParams = TRUE)
}
\arguments{
\item{obj}{TMB object}
}
\value{
Return estimated parameters and uncertainty
}
\description{
Title
}
\examples{

}
