% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_TMB_data.R
\name{create_TMB_data}
\alias{create_TMB_data}
\title{Prepare OM data in a data list suitabke for the TMB EM model}
\usage{
create_TMB_data(sim.data, df, history = FALSE)
}
\description{
Prepare OM data in a data list suitabke for the TMB EM model
}
