% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_TMB_data.R
\name{create_TMB_data}
\alias{create_TMB_data}
\title{Create the data for \link{TMB}}
\usage{
create_TMB_data(sim.data, df, history = FALSE)
}
\arguments{
\item{sim.data}{Operating model}

\item{df}{input parameters}

\item{history}{Use historical real data or OM data?}
}
\value{
A list of the data needed by \link{TMB}
}
\description{
Create the data for \link{TMB}
}
