% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_TMB_data.R
\name{create_TMB_data}
\alias{create_TMB_data}
\title{Create the data for \code{\link[TMB:MakeADFun]{TMB::MakeADFun()}}}
\usage{
create_TMB_data(sim_data = NULL, df = NULL, history = FALSE)
}
\arguments{
\item{sim_data}{Operating model}

\item{df}{input parameters}

\item{history}{Logical. If TRUE use historical data. If FALSE, use simulated OM data}
}
\value{
A list of the data needed by \code{\link[TMB:MakeADFun]{TMB::MakeADFun()}}
}
\description{
Create the data for \code{\link[TMB:MakeADFun]{TMB::MakeADFun()}}
}
