% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_TMB_data.R
\name{create_TMB_data}
\alias{create_TMB_data}
\title{Create the data for \code{\link[TMB:MakeADFun]{TMB::MakeADFun()}}}
\usage{
create_TMB_data(
  sim_data = NULL,
  df = NULL,
  ss_model = NULL,
  history = FALSE,
  sim_age_comps = TRUE
)
}
\arguments{
\item{sim_data}{Operating model}

\item{df}{input parameters}

\item{ss_model}{SS3 model output as created by \code{\link[=create_rds_file]{create_rds_file()}}
and loaded by \code{\link[=load_ss_model_from_rds]{load_ss_model_from_rds()}}}

\item{history}{Logical. If TRUE use historical data. If FALSE, use simulated OM data}

\item{sim_age_comps}{Logical. If TRUE, include simulated age comps (\code{age_survey} and \code{age_catch}),
If FALSE, leave them as-is}
}
\value{
A list of the data needed by \code{\link[TMB:MakeADFun]{TMB::MakeADFun()}}
}
\description{
Create the data for \code{\link[TMB:MakeADFun]{TMB::MakeADFun()}}
}
