% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_TMB_data.R
\name{create_tmb_data}
\alias{create_tmb_data}
\title{Create the data for \code{\link[TMB:MakeADFun]{TMB::MakeADFun()}}}
\usage{
create_tmb_data(om = NULL, ss_model = NULL, yr = NULL)
}
\arguments{
\item{om}{Operating model output returned by the \code{\link[=run_om]{run_om()}} function
and loaded by \code{\link[=load_ss_model_from_rds]{load_ss_model_from_rds()}}}

\item{ss_model}{A model input/output list representing the SS model as returned from
\code{\link[=load_ss_model_from_rds]{load_ss_model_from_rds()}}}

\item{yr}{The year to trim the data to on the end of the time series. Various objects
are populated into the future and must be trimmed}
}
\value{
A list of 2 elements: the data and parameter values needed by \code{\link[TMB:MakeADFun]{TMB::MakeADFun()}}
}
\description{
Create the data for \code{\link[TMB:MakeADFun]{TMB::MakeADFun()}}
}
\details{
This function mostly modified objects used in the OM which are
\link{data.frame} and \link{array} types to be \link{matrix} and \link{vector} types that are required
by TMB. It should be called on each iteration of the MSE loop before the call to
\code{\link[TMB:MakeADFun]{TMB::MakeADFun()}}.
}
