% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_TMB_data.R
\name{create_tmb_data}
\alias{create_tmb_data}
\title{Create the data for \code{\link[TMB:MakeADFun]{TMB::MakeADFun()}}}
\usage{
create_tmb_data(sim_data = NULL, df = NULL, ss_model = NULL)
}
\arguments{
\item{sim_data}{Operating model}

\item{df}{input parameters}

\item{ss_model}{SS3 model output as created by \code{\link[=create_rds_file]{create_rds_file()}}
and loaded by \code{\link[=load_ss_model_from_rds]{load_ss_model_from_rds()}}}
}
\value{
A list of 2 elements: the data and parameter values needed by \code{\link[TMB:MakeADFun]{TMB::MakeADFun()}}
}
\description{
Create the data for \code{\link[TMB:MakeADFun]{TMB::MakeADFun()}}
}
\details{
This function mostly modified objects used in the OM which are
\link{data.frame} and \link{array} types to be \link{matrix} and \link{vector} types that are required
by TMB. It should be called on each iteration of the MSE loop before the call to
\code{\link[TMB:MakeADFun]{TMB::MakeADFun()}}.
}
