% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/on_load.R
\name{.onLoad}
\alias{.onLoad}
\title{Load the dynamic library when the package is loaded}
\usage{
.onLoad(lib, pkg)
}
\arguments{
\item{lib}{The name of the dynamic link library}

\item{pkg}{The package name}
}
\value{
Nothing
}
\description{
Load the dynamic library when the package is loaded
}
\details{
If running \code{\link[devtools:check]{devtools::check()}}, be sure to call it as: \code{devtools::check(cran = FALSE)}
to avoid a major error about the DLL not being able to be found
}
