% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ss-model-loading.R
\name{extract_age_comps}
\alias{extract_age_comps}
\title{Extract the age comp estimates from the SS assessment model output into a format
required for input into the TMB assessment model}
\usage{
extract_age_comps(
  ss_model = NULL,
  age_comps_fleet = 1,
  s_yr = NULL,
  m_yr = NULL,
  age_comps_fill = -1,
  yr_col = "Yr",
  age_comps_fleet_col = "Fleet",
  ...
)
}
\arguments{
\item{ss_model}{SS model input/output as read in by \code{\link[=load_ss_model_from_rds]{load_ss_model_from_rds()}}}

\item{age_comps_fleet}{1 for fishery, 2 for survey}

\item{s_yr}{See \code{\link[=load_data_om]{load_data_om()}}}

\item{m_yr}{See \code{\link[=load_data_om]{load_data_om()}}}

\item{age_comps_fill}{Value to replace NAs in the table, it can also be NA, but not NULL}

\item{yr_col}{The name of the column in \code{ss_model$dat$agecomp} that contains
the year}

\item{age_comps_fleet_col}{The name of the column in \code{ss_model$dat$agecomp} that contains
the fleet code}

\item{...}{Arguments absorbed which are meant for other functions}
}
\value{
The matrix representing all years from \code{s_yr} to \code{m_yr} with data included
for years which are found in the \code{ss_model} data. Years not in the \code{ss_model} data
will be filled with the value of \code{age_comps_fill}
}
\description{
Extract the age comp estimates from the SS assessment model output into a format
required for input into the TMB assessment model
}
\details{
Proportions at age for each year are calculated and returned
}
