% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch-extra-mcmc.R
\name{fetch_extra_mcmc}
\alias{fetch_extra_mcmc}
\title{Create and return a list of stats to attach to the main model by
looking in the model's path for the report files.}
\usage{
fetch_extra_mcmc(model = NULL, ss_mcmc_quants = NULL, ...)
}
\arguments{
\item{model}{The model object as output by \code{\link[=load_ss_files]{load_ss_files()}}}

\item{ss_mcmc_quants}{Quantile probability values to use}

\item{...}{Absorb arguments meant for other functions}
}
\value{
A \link{list} of values ectracted and calculed from the extra MCMC runs
}
\description{
Create and return a list of stats to attach to the main model by
looking in the model's path for the report files.
}
