% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_plot_MSE.R
\name{fn_plot_MSE}
\alias{fn_plot_MSE}
\title{Produce several MSE plots}
\usage{
fn_plot_MSE(
  ls,
  sim.data,
  plotfolder = "results/Figs/newplots/",
  plotexp = TRUE,
  pidx = NA
)
}
\description{
Produce several MSE plots
}
