% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_plot_MSE.R
\name{fn_plot_MSE}
\alias{fn_plot_MSE}
\title{Title}
\usage{
fn_plot_MSE(
  ls,
  sim.data,
  plotfolder = "results/Figs/newplots/",
  plotexp = TRUE,
  pidx = NA
)
}
\arguments{
\item{ls}{list of MSE runs}

\item{sim.data}{baseline operating model}

\item{plotfolder}{folder to save figures}

\item{plotexp}{print figures to disk?}

\item{pidx}{order of runs in figures}
}
\value{

}
\description{
Title
}
\examples{

}
