% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fnMSE.R
\name{fnMSE}
\alias{fnMSE}
\title{Title}
\usage{
fnMSE(df, simyears = 30, TAC = 1, seeds = NA, nruns = 100)
}
\arguments{
\item{df}{parameters}

\item{simyears}{number of years to simulate}

\item{TAC}{which harvest control rule to use}

\item{seeds}{seed}

\item{nruns}{number of iterations}
}
\value{

}
\description{
Title
}
\examples{

fnMSE(df, nruns = 15) # Run MSE with 15 iterations

}
