% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_f.R
\name{get_f}
\alias{get_f}
\title{Calculate the fishing mortality based on catch.
See Methot & Wetzel 2013, Appendix A.}
\usage{
get_f(
  e_tmp = NULL,
  b_tmp = NULL,
  yr = yr,
  season = season,
  space = space,
  m_season = NULL,
  f_sel = NULL,
  n_tmp = NULL,
  wage_catch = NULL,
  method = "Hybrid"
)
}
\arguments{
\item{e_tmp}{Catch}

\item{b_tmp}{Biomass}

\item{m_season}{Natural mortality}

\item{f_sel}{Fisheries selectivity}

\item{n_tmp}{Numbers at age}

\item{wage_catch}{Weight at age}

\item{method}{Only 'Hybrid' implemented}
}
\value{
The fishing mortality value
}
\description{
Calculate the fishing mortality based on catch.
See Methot & Wetzel 2013, Appendix A.
}
