% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ref_point.R
\name{get_ref_point}
\alias{get_ref_point}
\title{Get reference points (TODO: improve docs on this function)}
\usage{
get_ref_point(
  pars,
  df,
  ssb_y,
  f_in = NA,
  n_end,
  tac = NULL,
  v_real = NA,
  space = 2,
  catch_floor = NULL,
  ...
)
}
\arguments{
\item{pars}{estimated parameters}

\item{df}{data frame of non-estimated parameters}

\item{ssb_y}{Spawning biomass}

\item{f_in}{Fishing mortality}

\item{n_end}{Numbers at age}

\item{tac}{which tac function to use?}

\item{v_real}{Vulnerable biomass in OM}

\item{space}{The space (area) to use the selectivity values for. Defaults to USA (2)}

\item{catch_floor}{The lowest catch to be allowed}

\item{...}{Absorb arguments meant for other functions}
}
\value{
A list of reference points
}
\description{
Get reference points (TODO: improve docs on this function)
}
