% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_select.R
\name{get_select}
\alias{get_select}
\title{Get a vector of selectivities of the same length as the \code{ages} vector
based on a range asked for}
\usage{
get_select(ages = NULL, p_sel = NULL, s_min = NULL, s_max = NULL, yr = 0)
}
\arguments{
\item{ages}{A vector of ages to calculate selectivity for}

\item{p_sel}{Log selectivity values from age \code{s_min} to \code{s_max}}

\item{s_min}{Minimum age}

\item{s_max}{Maximum age}
}
\value{
A vector of selectivity values of the same length as the \code{ages} vector
}
\description{
Get a vector of selectivities of the same length as the \code{ages} vector
based on a range asked for
}
