% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_weight_dataframes.R
\name{get_wa_dfs}
\alias{get_wa_dfs}
\title{Get the weight-at-X data for a particular year}
\usage{
get_wa_dfs(lst, yr)
}
\arguments{
\item{lst}{A list returned from \code{\link[=load_data_seasons]{load_data_seasons()}}}

\item{yr}{The year to fetch}
}
\value{
A list of 4 \link{data.frames}: \code{catch}, \code{surv}, \code{mid}, and \code{ssb}
}
\description{
Get the weight-at-X data for a particular year
}
