% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getF.R
\name{getF}
\alias{getF}
\title{Calculate the fishing mortality based on catch}
\usage{
getF(Etemp, Btemp, Mseason, Fsel, Ntemp, w_catch = w_catch, method = "Pope")
}
\arguments{
\item{Etemp}{Catch}

\item{Btemp}{Biomass}

\item{Mseason}{natural mortality}

\item{Fsel}{Fisheries selectivity}

\item{Ntemp}{Numbers at age}

\item{w_catch}{weight at age}

\item{method}{'Pope' or 'Hybrid' method?}
}
\value{

}
\description{
Calculate the fishing mortality based on catch
}
\examples{

}
