% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getParameters_csv.R
\name{getParameters_csv}
\alias{getParameters_csv}
\title{Title}
\usage{
getParameters_csv(trueparms = TRUE)
}
\arguments{
\item{trueparms}{true of random parameter initialization}
}
\value{

}
\description{
Title
}
\examples{
Export EM parameters for TMB based on csv files with parameters
}
