% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getParameters_csv.R
\name{getParameters_csv}
\alias{getParameters_csv}
\title{Export EM parameters for TMB based on csv files with parameters}
\usage{
getParameters_csv(trueparms = TRUE)
}
\description{
Export EM parameters for TMB based on csv files with parameters
}
