% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getParameters_OM.R
\name{getParameters_OM}
\alias{getParameters_OM}
\title{Create a list of parameters for TMB}
\usage{
getParameters_OM(trueparms = TRUE, mod = NA, df)
}
\arguments{
\item{trueparms}{should the model use the assesment parameters}

\item{mod}{r4ss object with a stock assessment}

\item{df}{list of parameters}
}
\value{

}
\description{
Create a list of parameters for TMB
}
