% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getParameters_ss.R
\name{getParameters_ss}
\alias{getParameters_ss}
\title{Export EM parameters for TMB based on csv files with parameters}
\usage{
getParameters_ss(trueparms = TRUE, mod = NA)
}
\description{
Export EM parameters for TMB based on csv files with parameters
}
