% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getParameters_ss.R
\name{getParameters_ss}
\alias{getParameters_ss}
\title{Load parameters from stock synthesis}
\usage{
getParameters_ss(mod = NA)
}
\arguments{
\item{mod}{}

\item{trueparms}{Should the parameters be the}
}
\value{
returns a list suitable for the TMB model
}
\description{
Load parameters from stock synthesis
}
\examples{
df <- getParameters_ss(mod) # check
Export EM parameters for TMB based on csv files with parameters
}
