% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getParms_noAge.R
\name{getParms_noage}
\alias{getParms_noage}
\title{Get a list of parameters for \code{\link[TMB:MakeADFun]{TMB::MakeADFun()}}}
\usage{
getParms_noage(hakedat)
}
\arguments{
\item{hakedat}{Hake data (TODO: where is this from? What is it's structure?)}
}
\value{
A list of parameters for \code{\link[TMB:MakeADFun]{TMB::MakeADFun()}}
}
\description{
Get a list of parameters for \code{\link[TMB:MakeADFun]{TMB::MakeADFun()}}
}
