% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getUncertainty.R
\name{getUncertainty}
\alias{getUncertainty}
\title{Calculate and organize the uncertainty on parameters and derived values}
\usage{
getUncertainty(name, data, sdrep)
}
\arguments{
\item{name}{name of the parameter or value to extract}

\item{data}{vector of data for organizing the years - remove later}

\item{sdrep}{list of standard deviations}
}
\value{
returns a data frame ready for plotting
}
\description{
Calculate and organize the uncertainty on parameters and derived values
}
