% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getUncertainty.R
\name{getUncertainty}
\alias{getUncertainty}
\title{Get uncertainty (TODO: Improve docs on this function)}
\usage{
getUncertainty(name, data, sdrep)
}
\arguments{
\item{name}{The parameter name}

\item{data}{Data}

\item{sdrep}{A \link{data.frame} of standard deviations}
}
\value{
A \link{data.frame} containing uncertainty
}
\description{
Get uncertainty (TODO: Improve docs on this function)
}
