% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hake_objectives.R
\name{hake_objectives}
\alias{hake_objectives}
\title{Title}
\usage{
hake_objectives(ls.MSE, SSB0, move = NA)
}
\arguments{
\item{ls.MSE}{list of MSE results}

\item{SSB0}{unfished biomass from OM}

\item{move}{is movement in the OM?}
}
\value{

}
\description{
Title
}
\examples{

}
