% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hake_objectives.R, R/hake_objectives_vs2.R
\name{hake_objectives}
\alias{hake_objectives}
\title{Title}
\usage{
hake_objectives(ls.MSE, SSB0, move = NA)

hake_objectives(ls.MSE, SSB0, move = NA)
}
\arguments{
\item{ls.MSE}{A list of individual MSE runs. Should contain Biomass, Catch, and df with parameters}

\item{SSB0}{Operating Model Spawning Stock B0 value}

\item{move}{is movement included?}
}
\value{


Nothing
}
\description{
Title

TODO: What this function does
}
\examples{

}
