% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hake_objectives.R
\name{hake_objectives}
\alias{hake_objectives}
\title{Hake objectives (TODO: Improve docs on this function)}
\usage{
hake_objectives(
  sim_data = NULL,
  short_term_yrs = NULL,
  long_term_yrs = NULL,
  quants = c(0.025, 0.05, 0.1, 0.25, 0.5, 0.75, 0.9, 0.95, 0.975),
  catch_multiplier = 1e-06,
  ...
)
}
\arguments{
\item{sim_data}{simulated data from the OM}

\item{short_term_yrs}{Years included in short term plots}

\item{long_term_yrs}{Years greater than this will be in long term plots}

\item{quants}{Quantiles to calculate for plotting output \link{data.frame}s}

\item{catch_multiplier}{Value to multiply all catch calculations by}

\item{...}{Absorbs extra parameters}
}
\value{
List of many outputs from the MSE
}
\description{
Hake objectives (TODO: Improve docs on this function)
}
