% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hake_objectives.R
\name{hake_objectives}
\alias{hake_objectives}
\title{Hake objectives (TODO: Improve docs on this function)}
\usage{
hake_objectives(
  lst = NULL,
  ssb0 = NULL,
  move = NA,
  short_term_yrs = 2018:2027,
  long_term_yrs = 2027,
  can.prop = 0.2488,
  us.prop = 0.7612
)
}
\arguments{
\item{lst}{list of MSE results}

\item{ssb0}{unfished biomass from OM}

\item{move}{Logical. Is movement in the OM?}

\item{short_term_yrs}{Years for short term plots}

\item{long_term_yrs}{Years greater than this will be in long term plots}
}
\value{
List of three: p.export, t.export, ls.season (TODO: explain better)
}
\description{
Hake objectives (TODO: Improve docs on this function)
}
