% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hake_violin.R
\name{hake_violin}
\alias{hake_violin}
\title{Violin plot of the MSE results}
\usage{
hake_violin(ls.MSE, SSB0, move = NA)
}
\arguments{
\item{ls.MSE}{list of MSE results}

\item{SSB0}{OM SSB0}

\item{move}{is movement included?}
}
\value{

}
\description{
Violin plot of the MSE results
}
