% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hake_violin.R
\name{hake_violin}
\alias{hake_violin}
\title{Violin plot of the MSE results}
\usage{
hake_violin(mse_output, om_output, lst_indicators, move = NA)
}
\arguments{
\item{mse_output}{MSE output of a single scenario in the format enforced in \code{\link[=setup_mse_plot_objects]{setup_mse_plot_objects()}}}

\item{om_output}{OM output of a single scenario in the format enforced in \code{\link[=setup_mse_plot_objects]{setup_mse_plot_objects()}}}

\item{lst_indicators}{Objectives output of a single scenario in the format enforced in \code{\link[=hake_objectives]{hake_objectives()}}}

\item{move}{Logical. If TRUE, stock movement is included}
}
\value{
A \link{data.frame} with columns SSB, SSB.10, SSB.40, Catch.short,
Catch.long, and AAV
}
\description{
Violin plot of the MSE results
}
