% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_agebased_model.R
\name{init_agebased_model}
\alias{init_agebased_model}
\title{Initialize values used in the age based model}
\usage{
init_agebased_model(df = NULL, lst = NULL)
}
\arguments{
\item{df}{A \link{list} as returned by \code{\link[=load_data_seasons]{load_data_seasons()}}}

\item{lst}{A \link{list} as returned by \code{\link[=setup_blank_om_objects]{setup_blank_om_objects()}}}
}
\value{
A \link{list} with the same structure os \code{lst}, but with a new element,
\code{n_init} added, and some values chaged for the initializeation of the model
}
\description{
Initialize values used in the age based model
}
