% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data_seasons.R
\name{load_data_seasons}
\alias{load_data_seasons}
\title{Load data frame containing all the parameters from the hake operating model}
\usage{
load_data_seasons(
  nseason = 4,
  nspace = 2,
  myear = 2018,
  movemaxinit = 0.35,
  movefiftyinit = 6,
  nsurvey = 2,
  logSDR = 1.4,
  bfuture = 0.5,
  moveout = 0.85,
  movesouth = 0.05,
  moveinit = NA,
  moveslope = 0.9,
  selectivity_change = 0,
  yr_future = 0,
  sel_hist = 1
)
}
\description{
Load data frame containing all the parameters from the hake operating model
}
