% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data_seasons.R
\name{load_data_seasons}
\alias{load_data_seasons}
\title{Prepare the data for the operating model}
\usage{
load_data_seasons(
  n_season = 4,
  season_names = NULL,
  n_space = 2,
  space_names = NULL,
  n_survey = 2,
  s_yr = 1966,
  m_yr = 2018,
  ages = 0:20,
  age_names = NULL,
  rdev_sd = 1.4,
  move_init = NULL,
  move_max_init = 0.35,
  move_fifty_init = 6,
  move_out = 0.85,
  move_south = 0.05,
  move_slope = 0.9,
  ages_no_move = c(0, 1),
  selectivity_change = 0,
  b_future = 0.5,
  yr_future = 0,
  sel_change_yr = 1991,
  sel_hist = TRUE,
  f_space = c(0.2612, 0.7388),
  catch_props_space_season = NULL
)
}
\arguments{
\item{n_season}{Number of seasons}

\item{season_names}{A vector of names for the seasons. Length must equal \code{n_season}}

\item{n_space}{Number of spatial areas}

\item{space_names}{A vector of names for the spaces. Length must equal \code{n_space}}

\item{n_survey}{Survey frequency}

\item{s_yr}{First year of historical simulations}

\item{m_yr}{Last year of historical simulations}

\item{ages}{A vector of ages}

\item{age_names}{A vector of names for the ages. Length must equal length of \code{ages}}

\item{rdev_sd}{Recruitment deviation Standard deviation}

\item{move_init}{Initial distribution of fish}

\item{move_max_init}{Maximum movement rate}

\item{move_fifty_init}{Age at 50 percent maximum movement rate}

\item{move_out}{Fraction of individuals that travel south in the last season}

\item{move_south}{Fraction of individuals that move south during the year}

\item{move_slope}{Slope of the movement function}

\item{ages_no_move}{Ages of fish which do not move in the movement model}

\item{selectivity_change}{Should selectivity change?}

\item{b_future}{Bias adjustment in the future}

\item{yr_future}{How many years into the future should there be stochastic values}

\item{sel_change_yr}{A year in which a selectivity change took place}

\item{sel_hist}{Use historical selectivity?}

\item{f_space}{The proportion of TAC given to each country. First value is Canada,
the second is the US}

\item{catch_props_space_season}{Proportion of catch to take by season and space}
}
\value{
A list of Parameters, Input parameters, Survey, Catch, and others
}
\description{
Prepare the data for the operating model
}
\examples{
\dontrun{
df <- load_data_seasons(n_season = 2, n_space = 2)
}
}
