% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data_ss.R
\name{load_data_ss}
\alias{load_data_ss}
\title{Set up a TMB data list from a SS object}
\usage{
load_data_ss(mod, sum_zero = 0)
}
\arguments{
\item{mod}{# SS3 object}

\item{sum_zero}{# Force the recruitment deviations to sum to zero?}
}
\value{
Returns a list for TMB
}
\description{
Set up a TMB data list from a SS object
}
