% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ss_model_loading.R
\name{load_ss_model_data}
\alias{load_ss_model_data}
\title{Load the SS model input and output data needed by this package in correct format}
\usage{
load_ss_model_data(
  s_min = 1,
  s_max = 6,
  s_min_survey = 2,
  s_max_survey = 6,
  weight_factor = 1000,
  n_space = 2,
  selex_fill_val = 1,
  ...
)
}
\arguments{
\item{s_min}{Minimum age in fishery selectivity}

\item{s_max}{Maximum age in fishery selectivity}

\item{s_min_survey}{Minimum age in survey selectivity}

\item{s_max_survey}{Maximum age in survey selectivity}

\item{weight_factor}{A factor to multiply and divide SSB values by}

\item{n_space}{The number of spaces (areas) in which fishing takes place
and to which a selectivity-at-age is to be set. See the value of
\code{p_sel_fish} in the returned list}

\item{selex_fill_val}{The selectivity value to fill in for missing
selectivity estimates, e.g. Used to fill in Canadian selectivity because the
estimation model only estimates one selectivity which is used for the USA}

\item{...}{Arguments to be passed to \code{\link[=load_ss_sel_parameters]{load_ss_sel_parameters()}}}
}
\value{
A \link{list} with objects to be used throughout the package code
}
\description{
Load the SS model input and output data needed by this package in correct format
}
