% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ss-model-loading.R
\name{load_ss_model_data}
\alias{load_ss_model_data}
\title{Load the SS model input and output data needed by this package in correct format}
\usage{
load_ss_model_data(ss_model, ...)
}
\arguments{
\item{ss_model}{SS3 model output as created by \code{\link[=create_rds_file]{create_rds_file()}}}

\item{...}{Arguments to be passed to \code{\link[=load_ss_sel_parameters]{load_ss_sel_parameters()}}}
}
\value{
A \link{list} with objects to be used throughout the package code
}
\description{
Load the SS model input and output data needed by this package in correct format
}
