% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ss_model_loading.R
\name{load_ss_sel_parameters}
\alias{load_ss_sel_parameters}
\title{Load a selectivity parameter estimates from the SS model output}
\usage{
load_ss_sel_parameters(
  ss_model = NULL,
  s_min = NULL,
  s_max = NULL,
  s_min_survey = NULL,
  s_max_survey = NULL,
  ...
)
}
\arguments{
\item{ss_model}{SS3 model output as created by \code{\link[=create_rds_file]{create_rds_file()}}}

\item{s_min}{Minimum age in fishery selectivity}

\item{s_max}{Maximum age in fishery selectivity}

\item{s_min_survey}{Minimum age in survey selectivity}

\item{s_max_survey}{Maximum age in survey selectivity}

\item{...}{Arguments absorbed which are meant for other functions}
}
\value{
A \link{matrix} of estimates, age, and source (fishery and survey coded as 1 and 2 respectively)
}
\description{
Load a selectivity parameter estimates from the SS model output
}
