% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pacifichakemse.R
\docType{package}
\name{pacifichakemse}
\alias{pacifichakemse}
\title{Pacific Hake MSE package}
\description{
Management strategy evaluation framework with a Spatial Operating Model.
}
\references{
PacifichakeMSE on GitHub: \url{https://github.com/nissandjac/PacifichakeMSE}
}
\author{
Nis S Jacobsen, Ian G. Taylor, Kristin Marshall, Aaron Berger, Chris Grandin
Package maintainer: Nis Jacobsen \href{mailto:nissandjac@gmail.com}{nissandjac@gmail.com}
}
