% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pacifichakemse.R
\docType{package}
\name{pacifichakemse}
\alias{pacifichakemse-package}
\alias{pacifichakemse}
\title{Pacific Hake MSE package}
\description{
Management strategy evaluation framework with a Spatial
Operating Model.

All of these imports are in alphabetical order, both packages and
functions within them
}
\references{
pacifichakemse on GitHub:
\url{https://github.com/pacific-hake/pacifichakemse}
}
\seealso{
Useful links:
\itemize{
  \item Report bugs at \url{https://github.com/pacific-hake/pacifichakemse/issues}
}

}
\author{
Chris J. Grandin, Nis S. Jacobsen, Ian G. Taylor, Kristin Marshall,
Aaron Berger
Package maintainer: Chris J. Grandin \href{mailto:chris.grandin@gmail.com}{chris.grandin@gmail.com}
}
