% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_aas.R
\name{plot_aa}
\alias{plot_aa}
\title{Plot Average Age in Survey (AAS) or Catch (AAC) by scenario}
\usage{
plot_aa(ps = NULL, type = NULL)
}
\arguments{
\item{ps}{A plot setup object as output by \code{\link[=setup_mse_plot_objects]{setup_mse_plot_objects()}}}

\item{type}{Which data type to plot, "survey" or "catch"}
}
\value{
A \code{\link[ggplot2:ggplot]{ggplot2::ggplot()}} object
}
\description{
Plot Average Age in Survey (AAS) or Catch (AAC) by scenario
}
