% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_catch_quota.R
\name{plot_catch_quota}
\alias{plot_catch_quota}
\title{Plot Catch/Quota by scenario}
\usage{
plot_catch_quota(ps = NULL)
}
\arguments{
\item{ps}{A plot setup object as output by \code{\link[=setup_mse_plot_objects]{setup_mse_plot_objects()}}}
}
\value{
A \code{\link[ggplot2:ggplot]{ggplot2::ggplot()}} object
}
\description{
Plot Catch/Quota by scenario
}
