% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_catch_quota.R
\name{plot_catch_quota}
\alias{plot_catch_quota}
\title{Plot Catch/Quota by scenario}
\usage{
plot_catch_quota(
  ps = NULL,
  ci = c(0.05, 0.95),
  yr_lim = c(NA_real_, NA_real_),
  y_lim = c(NA_real_, NA_real_)
)
}
\arguments{
\item{ps}{A plot setup object as output by \code{\link[=setup_mse_plot_objects]{setup_mse_plot_objects()}}}

\item{ci}{A vector of length two of the lower and upper credible interval values.}
}
\value{
A \code{\link[ggplot2:ggplot]{ggplot2::ggplot()}} object
}
\description{
Plot Catch/Quota by scenario
}
