% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_em_timeseries.R
\name{plot_em_timeseries}
\alias{plot_em_timeseries}
\title{Plot the mean of the year medians SSB for all Estimation models}
\usage{
plot_em_timeseries(
  ps,
  type = "ssb",
  ci = c(0.05, 0.95),
  ci_lines = TRUE,
  yr_lim = c(NA_real_, NA_real_),
  rev_scenarios = TRUE,
  legend_position = c(0.15, 0.92),
  free_y_scale = FALSE,
  ...
)
}
\arguments{
\item{ps}{A plot setup object as output by \code{\link[=setup_mse_plot_objects]{setup_mse_plot_objects()}}.}

\item{type}{One of 'ssb', 'ssb_ssb0', catch', 'aas', 'aac','aap'}

\item{ci}{A vector of length two of the lower and upper credible interval values.}

\item{ci_lines}{If TRUE, use dashed lines for the CI envelope. If FALSE, use a shaded ribbon.}

\item{yr_lim}{A vector of 2 for minimum and maximum yrs to show on the plot. If either are NA,
the limits of the data are used.}

\item{rev_scenarios}{If TRUE, reverse the order of the scenarios in the legend.}

\item{legend_position}{A vector of two - X and Y position for the legend. See \code{\link[ggplot2:theme]{ggplot2::theme()}}}

\item{free_y_scale}{See the \code{scales} argument of \code{\link[ggplot2:facet_wrap]{ggplot2::facet_wrap()}}}

\item{...}{Extra arguments to pass to \code{\link[=color_facet_backgrounds]{color_facet_backgrounds()}}}
}
\value{
A \code{\link[ggplot2:ggplot]{ggplot2::ggplot()}} object
}
\description{
Plot the mean of the year medians SSB for all Estimation models
}
