% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_exploitation_rate.R
\name{plot_exploitation_rate}
\alias{plot_exploitation_rate}
\title{Plot the exploitation rate (F0) by country and scenario}
\usage{
plot_exploitation_rate(ps = NULL, ci = c(0.05, 0.95))
}
\arguments{
\item{ps}{A plot setup object as output by \code{\link[=setup_mse_plot_objects]{setup_mse_plot_objects()}}}

\item{ci}{A vector of length two of the lower and upper credible interval values.}
}
\value{
A \code{\link[ggplot2:ggplot]{ggplot2::ggplot()}} object
}
\description{
Plot the exploitation rate (F0) by country and scenario
}
