% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_ssb.R
\name{plot_ssb}
\alias{plot_ssb}
\title{Plot Spawning Stock Biomass (SSB) by scenario and country}
\usage{
plot_ssb(ps = NULL, ci = c(0.05, 0.95))
}
\arguments{
\item{ps}{A plot setup object as output by \code{\link[=setup_mse_plot_objects]{setup_mse_plot_objects()}}.}

\item{ci}{A vector of length two of the lower and upper credible interval values.}
}
\value{
A \code{\link[ggplot2:ggplot]{ggplot2::ggplot()}} object
}
\description{
Plot Spawning Stock Biomass (SSB) by scenario and country
}
