% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_ssb_ssb0.R
\name{plot_ssb_ssb0}
\alias{plot_ssb_ssb0}
\title{Plot total SSB/SSB0 by scenario}
\usage{
plot_ssb_ssb0(ps = NULL, ci = c(0.05, 0.95), yr_lim = c(NA_real_, NA_real_))
}
\arguments{
\item{ps}{A plot setup object as output by \code{\link[=setup_mse_plot_objects]{setup_mse_plot_objects()}}}

\item{ci}{A vector of length two of the lower and upper credible interval values.}
}
\value{
A \code{\link[ggplot2:ggplot]{ggplot2::ggplot()}} object
}
\description{
Plot total SSB/SSB0 by scenario
}
