% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_standard_error.R
\name{plot_standard_error}
\alias{plot_standard_error}
\title{Plot Standard Error of the Spawning Stock Biomass between the Operating model and
the Estimation model}
\usage{
plot_standard_error(ps = NULL, quants = c(0.05, 0.95))
}
\arguments{
\item{ps}{A plot setup object as output by \code{\link[=setup_mse_plot_objects]{setup_mse_plot_objects()}}}

\item{quants}{A vector of quantiles to pass to \code{\link[stats:quantile]{stats::quantile()}}. Median is also
calculated}
}
\value{
A \code{\link[ggplot2:ggplot]{ggplot2::ggplot()}} object
}
\description{
Plot Standard Error of the Spawning Stock Biomass between the Operating model and
the Estimation model
}
