% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_tac_estimated_vs_realized.R
\name{plot_tac_vs_hcr}
\alias{plot_tac_vs_hcr}
\title{Plot estimated vs realized TAC (called Catch in plot) by harvest control rule
value}
\usage{
plot_tac_vs_hcr(df)
}
\arguments{
\item{df}{A \link{data.frame} as output by \code{\link[=calc_tac_est_vs_real]{calc_tac_est_vs_real()}}}
}
\value{
A \code{\link[ggplot2:ggplot]{ggplot2::ggplot()}} object
}
\description{
Plot estimated vs realized TAC (called Catch in plot) by harvest control rule
value
}
