% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_violins.R
\name{plot_violins}
\alias{plot_violins}
\title{Plot faceted violin plots for scenarios}
\usage{
plot_violins(ps, quants = c(0.05, 0.95), min_yr = 2020)
}
\arguments{
\item{ps}{A plot setup object as output by \code{\link[=setup_mse_plot_objects]{setup_mse_plot_objects()}}}
}
\value{
A \code{\link[ggplot2:ggplot]{ggplot2::ggplot()}} object
}
\description{
Plot faceted violin plots for scenarios
}
