% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotMSE.R
\name{plotMSE}
\alias{plotMSE}
\title{Plot MSE results (TODO: Improve docs on this function)}
\usage{
plotMSE(
  results_dir = NULL,
  figures_root_dir = here("figures"),
  figures_dir = NULL,
  plotnames = NULL,
  plotexp = FALSE,
  pidx = NA
)
}
\arguments{
\item{results_dir}{A directory name where results from \code{\link[=run_mses]{run_mses()}} resides. Only files
in this directory that end in '.rds' and contain 'MSE' will be loaded}

\item{figures_root_dir}{The root directory for figures. Scenario figure types will be
subdirectories called \code{figures_dir}.}

\item{figures_dir}{The directory to place the figures in.}

\item{plotnames}{Names for the plots}

\item{plotexp}{See \code{\link[=fn_plot_MSE]{fn_plot_MSE()}}}

\item{pidx}{See \code{\link[=fn_plot_MSE]{fn_plot_MSE()}}}
}
\value{
Nothing
}
\description{
Plot MSE results (TODO: Improve docs on this function)
}
