% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotSelectivity.R
\name{plotSelectivity}
\alias{plotSelectivity}
\title{Function for plotting age specific fisheries selectivity}
\usage{
plotSelectivity(df.new, parameters)
}
\description{
Function for plotting age specific fisheries selectivity
}
