% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotSelectivity.R
\name{plotSelectivity}
\alias{plotSelectivity}
\title{Title}
\usage{
plotSelectivity(df.new, parameters)
}
\arguments{
\item{df.new}{parameters and life history values}

\item{parameters}{estimates parameters}
}
\value{

}
\description{
Title
}
\examples{

}
