% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotViolin.R
\name{plotViolin}
\alias{plotViolin}
\title{Title}
\usage{
plotViolin(data, cols)
}
\arguments{
\item{data}{name of data set to be loaded}

\item{cols}{colors of the individual violins}
}
\value{

}
\description{
Title
}
\examples{
colors = c('black','green')
plotViolin(folder, colors) # plots black and green data

}
