% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df_MSE.R
\name{processMSE}
\alias{processMSE}
\title{Turn whole list into a concanated polished data frame}
\usage{
processMSE(
  df.MSE,
  id,
  idx = 1,
  spacenames = c("CAN", "USA"),
  runs = 100,
  nspace = 2
)
}
\arguments{
\item{df.MSE}{MSE that contains the data}

\item{id}{object to extract}

\item{idx}{which indexes to save to}

\item{spacenames}{naming convention for spatial areas}

\item{runs}{number of runs in MSE}

\item{nspace}{number of spaces}
}
\value{
returns a concanated data frame
}
\description{
Turn whole list into a concanated polished data frame
}
\examples{
processMSE(test) # not run

}
