% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_mse_scenario.R
\name{run_mse_scenario}
\alias{run_mse_scenario}
\title{Run a single MSE scenario}
\usage{
run_mse_scenario(
  om = NULL,
  random_seed = NULL,
  n_sim_yrs = NULL,
  tac = NULL,
  c_increase = 0,
  m_increase = 0,
  sel_change = 0,
  ...
)
}
\arguments{
\item{om}{List as output by \code{\link[=load_data_om]{load_data_om()}}}

\item{random_seed}{Seed for running the OM}

\item{n_sim_yrs}{Number of years to simulate}

\item{c_increase}{Increase in max movement}

\item{m_increase}{Decrease of spawners returning south}

\item{sel_change}{Time varying selectivity}

\item{...}{Absorb arguments intended for other functions}

\item{ss_model}{Output from \code{\link[=load_ss_model_data]{load_ss_model_data()}}}
}
\value{
A list of length 3: The MSE output, the OM output, and the EM output
}
\description{
Run a single MSE scenario
}
