% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_mse_scenario.R
\name{run_mse_scenario}
\alias{run_mse_scenario}
\title{Run a single MSE scenario}
\usage{
run_mse_scenario(
  om = NULL,
  random_seed = NULL,
  n_sim_yrs = NULL,
  tac = NULL,
  attain = c(1, 1),
  c_increase = 0,
  m_increase = 0,
  sel_change = 0,
  save_all_em = FALSE,
  ...
)
}
\arguments{
\item{om}{List as output by \code{\link[=load_data_om]{load_data_om()}}}

\item{random_seed}{Seed for running the OM}

\item{n_sim_yrs}{Number of years to simulate}

\item{tac}{The tac vector of length 1 or 2 to apply. e.g. c(a, b): the new catch
in the OM is c_new * b + a. If instead of a 2-element vector, a single value
is given, the expanded catch in the OM will be c_new * 0.5 unless below \code{catch_floor} in
which case it will be c_new = \code{catch_floor}.}

\item{attain}{The attainment vector of length 2, in the order Canada, US. These are proportions
of the catch to take.}

\item{c_increase}{Increase in max movement}

\item{m_increase}{Decrease of spawners returning south}

\item{sel_change}{Time varying selectivity}

\item{save_all_em}{If TRUE, the EM output will be saved for every year. If FALSE, it will be
saved for the last simulation year only.}

\item{...}{Absorb arguments intended for other functions}
}
\value{
A list of length 3: The MSE output, the OM output, and the EM output
}
\description{
Run a single MSE scenario
}
