% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_mses.R
\name{run_mses}
\alias{run_mses}
\title{Run the MSE for a number of runs for any scenarios}
\usage{
run_mses(
  ss_extdata_dir = NULL,
  nruns = 10,
  simyears = NULL,
  fns = NULL,
  tacs = NULL,
  cincreases = 0,
  mincreases = 0,
  sel_changes = 0,
  nsurveys = NULL,
  om_params_seed = 12345,
  results_root_dir = "results",
  results_dir = "default",
  ...
)
}
\arguments{
\item{ss_extdata_dir}{A directory in the PacifichakeMSE package extdata directory
which contains SS3 model run output}

\item{nruns}{Then number of runs to do for each simulation}

\item{simyears}{The number of years to simulate into the future}

\item{fns}{A vector of file names for the scenarios (.rds files). .rds extension is optional}

\item{tacs}{A vector of TAC values to be passed to the \code{\link[=run_multiple_MSEs]{run_multiple_MSEs()}} function, in the same
order as the \code{fns} file names, or a single value}

\item{cincreases}{A vector of values to be passed to the \code{\link[=run_multiple_MSEs]{run_multiple_MSEs()}} function, in the same
order as the \code{fns} file names, or a single value}

\item{mincreases}{A vector of values to be passed to the \code{\link[=run_multiple_MSEs]{run_multiple_MSEs()}} function, in the same
order as the \code{fns} file names, or a single value}

\item{sel_changes}{A vector of values to be passed to the \code{\link[=run_multiple_MSEs]{run_multiple_MSEs()}} function, in the same
order as the \code{fns} file names, or a single value}

\item{nsurveys}{The number of surveys for each run. This must be a vector of the same length as \code{fns} or \code{NULL}.
If \code{NULL}, nothing will be changed}

\item{om_params_seed}{A seed value to use when calling the \code{\link[=run.agebased.true.catch]{run.agebased.true.catch()}} function}

\item{results_root_dir}{The name of the results root directory (relative to \code{\link[here:here]{here::here()}})}

\item{results_dir}{The name of the results directory (relative to \link[here:here(results_root_dir)]{here::here(results_root_dir)})}

\item{...}{Arguments passed to \code{\link[=load_data_seasons]{load_data_seasons()}}}
}
\value{
Nothing
}
\description{
Run the MSE for a number of runs for any scenarios
}
\details{
Saves the output of the MSE runs to the files specified in \code{fns}, in the directories
specified in \code{results_root_dir} and \code{results_dir}
}
