% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_mses.R
\name{run_mses}
\alias{run_mses}
\title{Run the MSE for a number of runs for any scenarios}
\usage{
run_mses(
  n_runs = 10,
  n_sim_yrs = NULL,
  fns = NULL,
  plot_names = NULL,
  tacs = c(0, 1),
  attains = c(1, 1),
  c_increases = 0,
  m_increases = 0,
  sel_changes = 0,
  n_surveys = 2,
  b_futures = 0.5,
  random_seed = 12345,
  single_seed = NULL,
  results_root_dir = here("results"),
  results_dir = here("results", "default"),
  catch_floor = NULL,
  hcr_apply = FALSE,
  ...
)
}
\arguments{
\item{n_runs}{Then number of runs to do for each simulation}

\item{n_sim_yrs}{The number of years to simulate into the future}

\item{fns}{A vector of file names for the scenarios (.rds files). .rds extension is optional}

\item{plot_names}{A vector of strings to use for the scenarios later when plotting. Must either be
\code{NULL} or the same length as \code{fns}}

\item{tacs}{A vector of TAC values to be passed to the \code{\link[=run_mse_scenario]{run_mse_scenario()}} function, in the same
order as the \code{fns} file names, or a single value}

\item{attains}{A vector of 2, in the order Canada, US for proportion of catch to include}

\item{c_increases}{Increase in max movement. A vector of values to be passed to the \code{\link[=run_mse_scenario]{run_mse_scenario()}}
function, in the same order as the \code{fns} file names, or a single value which will be used for all scenarios}

\item{m_increases}{Decrease of spawners returning south. A vector of values to be passed to the
\code{\link[=run_mse_scenario]{run_mse_scenario()}} function, in the same order as the \code{fns} file names, or a single value which
will be used for all scenarios}

\item{sel_changes}{Selectivity scenario change type. A vector of values to be passed to the
\code{\link[=run_mse_scenario]{run_mse_scenario()}} function, in the same order as the \code{fns} file names, or a single value which will
be used for all scenarios}

\item{n_surveys}{The number of surveys for each run. This must be a vector of the same length as \code{fns} or \code{NULL}.
If \code{NULL}, 2 will be used for every scenario}

\item{b_futures}{A vector of values to be passed to the \code{\link[=run_mse_scenario]{run_mse_scenario()}} function for bias adjustment into
the future, in the same order as the \code{fns} file names, or a single value which will be used for all scenarios}

\item{random_seed}{A seed value to use when calling for all random functions}

\item{single_seed}{If NULL, ignore. If a number, use that as a seed to run a single run of the MSE. User for testing.}

\item{results_root_dir}{The results root directory}

\item{results_dir}{The results directory}

\item{catch_floor}{The lowest value to allow catch to drop to when applying the tac rule for the catch floor}

\item{...}{Arguments passed to \code{\link[=load_data_om]{load_data_om()}}}
}
\value{
Nothing
}
\description{
Run the MSE for a number of runs for any scenarios
}
\details{
Saves the output of the MSE runs to the files specified in \code{fns},
in the directories specified in \code{results_root_dir} and \code{results_dir}
}
