% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_multiple_MSEs.R
\name{run_multiple_MSEs}
\alias{run_multiple_MSEs}
\title{Iterate the pacific hake MSE}
\usage{
run_multiple_MSEs(
  simyears = NULL,
  seeds = 12345,
  TAC = 1,
  df = NA,
  cincrease = 0,
  mincrease = 0,
  sel_change = 0
)
}
\arguments{
\item{simyears}{number of years to simulate}

\item{seeds}{set the seed}

\item{TAC}{Which harvest control rule should the model use}

\item{df}{data frame of parameters}

\item{cincrease}{increase in max movement}

\item{mincrease}{decrease of spawners returning south}

\item{sel_change}{time varying selectivity}
}
\value{

}
\description{
Iterate the pacific hake MSE
}
