% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_multiple_MSEs.R
\name{run_multiple_MSEs}
\alias{run_multiple_MSEs}
\title{Run/Iterate the Pacific hake MSE}
\usage{
run_multiple_MSEs(
  df = NULL,
  simyears = NULL,
  seed = 12345,
  tac = 1,
  cincrease = 0,
  mincrease = 0,
  sel_change = 0,
  ...
)
}
\arguments{
\item{df}{Data frame of parameters as output by \code{\link[=load_data_seasons]{load_data_seasons()}}}

\item{simyears}{Number of years to simulate}

\item{seed}{The random number seed to use}

\item{tac}{Which harvest control rule should the model use}

\item{cincrease}{Increase in max movement}

\item{mincrease}{Decrease of spawners returning south}

\item{sel_change}{Time varying selectivity}

\item{...}{Absorb arguments intended for other functions}
}
\value{
A list of Catch, Catch.quota, SSB, SSB.mid, SSB.hes, Survey.om
F0, parms, N, converge, ams, amc, V
}
\description{
Run/Iterate the Pacific hake MSE
}
