% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_multiple_MSEs.R
\name{run_multiple_MSEs}
\alias{run_multiple_MSEs}
\title{Run/Iterate the Pacific hake MSE}
\usage{
run_multiple_MSEs(
  df = NULL,
  ss_model = NULL,
  sim_data = NULL,
  om_params_seed = 12345,
  n_sim_yrs = NULL,
  seed = 12345,
  tac = 1,
  c_increase = 0,
  m_increase = 0,
  sel_change = 0,
  ...
)
}
\arguments{
\item{df}{Data frame of parameters as output by \code{\link[=load_data_seasons]{load_data_seasons()}}}

\item{ss_model}{SS3 model output as created by \code{\link[=create_rds_file]{create_rds_file()}}
and loaded by \code{\link[=load_ss_model_from_rds]{load_ss_model_from_rds()}}}

\item{sim_data}{Operating model as created by \code{\link[=run_agebased_true_catch]{run_agebased_true_catch()}}}

\item{om_params_seed}{Seed for running the OM if it needs to be run (if \code{sim_data}
is \code{NULL})}

\item{n_sim_yrs}{Number of years to simulate}

\item{seed}{The random number seed to use for the MSE run}

\item{tac}{Which harvest control rule should the model use}

\item{c_increase}{Increase in max movement}

\item{m_increase}{Decrease of spawners returning south}

\item{sel_change}{Time varying selectivity}

\item{...}{Absorb arguments intended for other functions}
}
\value{
A list of Catch, Catch.quota, SSB, SSB.mid, SSB.hes, Survey.om
F0, parms, N, converge, ams, amc, V
}
\description{
Run/Iterate the Pacific hake MSE
}
