% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_multiple_OMs.R
\name{run_multiple_OMs}
\alias{run_multiple_OMs}
\title{Title}
\usage{
run_multiple_OMs(
  simyears = 30,
  seed = 12345,
  df,
  Catchin,
  cincrease = 0,
  mincrease = 0,
  sel_change = 0
)
}
\arguments{
\item{simyears}{number of years to simulate}

\item{seed}{set the seed}

\item{df}{data frame of parameters}

\item{Catchin}{catches in the future}

\item{cincrease}{increase in movement due to climate change}

\item{mincrease}{increase in the number of returning spawners}

\item{sel_change}{should selectivity be time varying?}
}
\value{
returns a list of results
 @export
}
\description{
Title
}
\examples{
run_multiple_OMs(df,100) Catch is 100 kg in the future

}
