% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_om.R
\name{run_om}
\alias{run_om}
\title{Run the Operating Model, which is an age based model
conditioned on true catch}
\usage{
run_om(df = NULL, om_objs = NULL, ...)
}
\arguments{
\item{df}{data frame of parameters and life history values}

\item{om_objs}{A \link{list} of the OM objects (arrays and matrices) for holding the
OM data. This list is what is returned by \code{\link[=setup_blank_om_objects]{setup_blank_om_objects()}}}

\item{...}{Arguments passed to \code{\link[=run_year_loop_om]{run_year_loop_om()}}}
}
\value{
A list of model outputs
}
\description{
Run the Operating Model, which is an age based model
conditioned on true catch
}
