% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_om.R
\name{run_om}
\alias{run_om}
\title{Run the Operating Model, which is an age based model
conditioned on true catch}
\usage{
run_om(df = NULL, seed = 100, n_sim_yrs = NULL, ...)
}
\arguments{
\item{df}{data frame of parameters and life history values}

\item{seed}{seed for survey error and recruitment deviations}

\item{n_sim_yrs}{The number of years to simulate into the future.}

\item{...}{Arguments passed to \code{\link[=run_year_loop_om]{run_year_loop_om()}}}
}
\value{
A list of model outputs (TODO)
}
\description{
Run the Operating Model, which is an age based model
conditioned on true catch
}
\examples{
\dontrun{
run_om(df)
}
}
