% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_season_loop_om.R
\name{run_season_loop_om}
\alias{run_season_loop_om}
\title{Run the Operating model for all seasons and spaces in one year}
\usage{
run_season_loop_om(
  df,
  lst,
  yr,
  yr_ind,
  m_season,
  pope_mul = 0.5,
  verbose = TRUE,
  ...
)
}
\arguments{
\item{df}{See \code{\link[=run_om]{run_om()}}}

\item{lst}{See \code{\link[=run_om]{run_om()}}}

\item{yr}{The year to run the operating model for}

\item{yr_ind}{The index of \code{yr} in the \code{df$yrs} vector}

\item{m_season}{A vector of natural mortality-at-age}

\item{pope_mul}{Multiplier used in Pope's method}

\item{verbose}{Print the loop information to the console}

\item{...}{Absorbs additional arguments meant for other functions}
}
\value{
A modified version of \code{lst} with the current data for \code{yr} populated
in all it's arrays and other objects
}
\description{
Run the Operating model for all seasons and spaces in one year
}
