% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_year_loop_om.R
\name{run_year_loop_om}
\alias{run_year_loop_om}
\title{Run the Operating model for all years}
\usage{
run_year_loop_om(df = NULL, lst = NULL, verbose = TRUE, ...)
}
\arguments{
\item{df}{See \code{\link[=run_agebased_true_catch]{run_agebased_true_catch()}}}

\item{lst}{See \code{\link[=run_agebased_true_catch]{run_agebased_true_catch()}}}

\item{...}{Additional arguments to be passed to \code{\link[=run_season_loop_om]{run_season_loop_om()}}}
}
\value{
A list
}
\description{
Run the Operating model for all years
}
