% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{setup_blank_om_objects}
\alias{setup_blank_om_objects}
\title{Construct blank objects for Operating model outputs}
\usage{
setup_blank_om_objects(
  yrs = NULL,
  ages = NULL,
  age_plus_grp = NULL,
  max_surv_age = NULL,
  n_space = NULL,
  n_season = NULL
)
}
\arguments{
\item{yrs}{A vector of years}

\item{ages}{A vector of ages}

\item{age_plus_grp}{Age for age plus group}

\item{max_surv_age}{The maximum age in the survey age comp data}

\item{n_space}{The number of spaces in the movement model}

\item{n_season}{The number of seasons in the movement model}
}
\value{
A list of the blank OM outputs
}
\description{
Construct blank objects for Operating model outputs
}
