% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_om_data.R
\name{update_om_data}
\alias{update_om_data}
\title{Update the OM data for a new year}
\usage{
update_om_data(
  sim_data = NULL,
  df = NULL,
  ss_model = NULL,
  sim_age_comps = TRUE
)
}
\arguments{
\item{sim_data}{Operating model}

\item{df}{Input data and parameters}

\item{ss_model}{SS3 model output as created by \code{\link[=create_rds_file]{create_rds_file()}}
and loaded by \code{\link[=load_ss_model_from_rds]{load_ss_model_from_rds()}}}

\item{sim_age_comps}{Logical. If TRUE, include simulated age comps (\code{age_survey} and \code{age_catch}),
If FALSE, leave them as-is}
}
\value{
A list of the data needed by \code{\link[TMB:MakeADFun]{TMB::MakeADFun()}}
}
\description{
Update the OM data for a new year
}
