% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_om_data.R
\name{update_om_data}
\alias{update_om_data}
\title{Update the OM data for a new year}
\usage{
update_om_data(
  yr = NULL,
  om = NULL,
  yr_survey_sims = NULL,
  f_new = NULL,
  c_increase = NULL,
  m_increase = NULL,
  sel_change = NULL,
  wage_only = FALSE,
  ...
)
}
\arguments{
\item{yr}{The year to add}

\item{om}{Operating model output as returned by \code{\link[=run_om]{run_om()}}}

\item{yr_survey_sims}{Years in which a survey takes place}

\item{f_new}{The output of the \code{\link[=get_ref_point]{get_ref_point()}} function from the previous year}

\item{c_increase}{Increase in max movement}

\item{m_increase}{Decrease of spawners returning south}

\item{sel_change}{Time varying selectivity}

\item{wage_only}{Only update the wage_ \link{data.frame}s, nothing else}

\item{...}{Absorb unused parameters
return a random normal using mean = 0 and sd = exp(om$rdev_sd)}
}
\value{
A list of the data needed by \code{\link[TMB:MakeADFun]{TMB::MakeADFun()}}
}
\description{
Update the OM data for a new year
}
