% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_om_data.R
\name{update_om_data}
\alias{update_om_data}
\title{Update the OM data for a new year}
\usage{
update_om_data(
  df = NULL,
  sim_data = NULL,
  yr = NULL,
  yr_ind = NULL,
  yr_survey_sims = NULL,
  f_new = NULL,
  c_increase = NULL,
  m_increase = NULL,
  sel_change = NULL,
  wage_only = FALSE,
  zero_rdevs = TRUE
)
}
\arguments{
\item{df}{Input data as created by \code{\link[=create_tmb_data]{create_tmb_data()}}}

\item{sim_data}{Operating model as created by \code{\link[=run_om]{run_om()}}}

\item{yr}{The year to add}

\item{yr_ind}{Index of the year to add}

\item{yr_survey_sims}{Years in which a survey takes place}

\item{f_new}{The output of the \code{\link[=get_ref_point]{get_ref_point()}} function from the previous year}

\item{c_increase}{Increase in max movement}

\item{m_increase}{Decrease of spawners returning south}

\item{sel_change}{Time varying selectivity}

\item{wage_only}{Only update the wage_ \link{data.frame}s, nothing else}

\item{zero_rdevs}{Logical. If TRUE, return recruitment devs (r_in) of 0. If FALSE,
return a random normal using mean = 0 and sd = exp(df$rdev_sd)}
}
\value{
A list of the data needed by \code{\link[TMB:MakeADFun]{TMB::MakeADFun()}}
}
\description{
Update the OM data for a new year
}
